==> 01_Spatial_Analysis.Rmd <==
---
title: "01_Spatial_Analysis"
author: "Matt"
date: "9/10/2020"
output: html_document
---

This document will outline the spatial analyses used in generating figures for my MSc. This code should be reproducible and should work into the future as long as the functions stay the same. First, load libraries and initialize folders.

```{r Load packages and set directories}

suppressMessages(suppressWarnings({
  dt <- "devtools"[!("devtools" %in% installed.packages()[, "Package"])]
  if(length(dt)) install.packages("devtools")
  if(!"devtools" %in% rownames(installed.packages())) install.packages("devtools")
  ls <- c("tidyverse", "bcmaps", "bcdata", "sf", "terra", "raster", "mapview")
  new_packages <- ls[!(ls %in% installed.packages()[, "Package"])]
  if(length(new_packages)) install.packages(new_packages)
  sapply(ls, library, character.only = TRUE, quietly = TRUE)[0] 
  rm(ls, new_packages)}))

data_dir <- file.path("./01_Spatial_Data/data")
ldb_maps_dir <- file.path("./01_Spatial_Data/layers/ldb")
bc_maps_dir <- file.path("./01_Spatial_Data/layers/bc")
results_dir <- file.path("./01_Spatial_Data/results")
dir.create(ldb_maps_dir, showWarnings = FALSE, recursive = TRUE)
dir.create(bc_maps_dir, showWarnings = FALSE, recursive = TRUE)
dir.create(results_dir, showWarnings = FALSE)

```

Next, generate a study area of Lac Du Bois from the publicly available BC Parks, Ecological Reserves, and Protected Areas dataset

```{r Lac Du Bois park}

ldb_aoi <- bcdata::bcdc_query_geodata("1130248f-f1a3-4956-8b2e-38d29d3e4af7", crs = 3005) %>% 
  bcdata::filter(PROTECTED_LANDS_NAME %in% c(
    "LAC DU BOIS GRASSLANDS PROTECTED AREA", 
    "MCQUEEN CREEK ECOLOGICAL RESERVE",
    "TRANQUILLE ECOLOGICAL RESERVE")) %>% 
  bcdata::collect() %>% 
  sf::st_union()

ldb_area <- units::set_units(st_area(ldb_aoi), km^2)

st_write(ldb_aoi, file.path(ldb_maps_dir, "aoi.gpkg"), delete_layer = TRUE, quiet = TRUE)

```

Next, do some BEC work. We want to see the BEC zones within Lac Du Bois park. 
NOTE: The coding here forces a new BEC polygon to be downloaded every time it is ran. To prevent that, change to "force = FALSE", though it doesn't take overly long anyways if you're patient!

```{r BEC Manipulation}

# Lac Du Bois BEC zones and subzones
ldb_bec <- bcmaps::bec(force = TRUE) %>% 
  sf::st_set_agr("constant") %>% 
  sf::st_intersection(ldb_aoi) %>% 
  group_by(ZONE, MAP_LABEL) %>% 
  summarise(do_union = TRUE, .groups = "drop") %>% 
  dplyr::rename(SUBZONE = MAP_LABEL) %>% 
  dplyr::mutate(bgc_area = units::set_units(st_area(.), km^2)) %>% 
  st_cast("MULTIPOLYGON") %>% 
  sf::st_set_agr("constant")

ldb_bec_ng <- st_drop_geometry(ldb_bec)
ldb_bgc_area <- sum(ldb_bec$bgc_area)

st_write(ldb_bec, file.path(ldb_maps_dir, "bec.gpkg"), delete_layer = TRUE, quiet = TRUE)

```

In this next section, we are going to use the IAPP (Invasive Alien Plant Program) database to find the spotted knapweed polygons in Lac Du Bois. 

```{r LDB IAPP plant sites}

# DATA DOWNLOAD:
# Download invasive plant boundaries from IAPP within the Lac Du Bois area and
# remove all polygons that are "ZZ" ONLY - polygons mixed with "ZZ" are okay. 
# ZZ stands for "no plants"
ldb_iapp <- bcdata::bcdc_query_geodata("10ecf9ad-1555-4043-834a-f5d24a506d59", crs = 3005) %>% 
  bcdata::filter(bcdata::BBOX(sf::st_bbox(ldb_aoi), crs = 3005)) %>% 
  collect() %>% 
  dplyr::filter(MAP_LABEL != "ZZ") %>% 
  sf::st_set_agr("constant") %>% 
  sf::st_intersection(ldb_aoi) %>% 
  sf::st_cast("MULTIPOLYGON") %>% 
  sf::st_set_agr("constant")

# Get the official plant list from the IAPP, it has species names etc.
curl::curl_download(
  "https://www2.gov.bc.ca/assets/gov/environment/plants-animals-and-ecosystems/invasive-species/iapp-resources/iapp_species_list_and_map_codes_feb2020.xlsx", 
  destfile = file.path(data_dir, "iapp_species_list_and_map_codes.xlsx"))

# Read in downloaded table
bc_iapp_plant_table <- readxl::read_xlsx(
  file.path(data_dir, "iapp_species_list_and_map_codes.xlsx"), col_names = TRUE) %>% 
  rename_with(make.names) %>% 
  dplyr::arrange(Map.Symbol) %>% 
  dplyr::rename(Map.Label = Map.Symbol) %>% 
  dplyr::filter(Map.Label != "ZZ")

# Extract vector of plant codes for Lac Du Bois only
iapp_plants <- bc_iapp_plant_table$Map.Label
ldb_iapp_plants <- sapply(iapp_plants, grep, ldb_iapp$MAP_LABEL) %>% 
  compact() %>% names()

# SPATIAL CALCULATIONS: Must start with total, then filter to species levels
# Create intersection with BGC subzone layer
ldb_all_sp_invaded_subzone <- st_intersection(ldb_iapp, ldb_bec)

# Summary of the full IAPP dataset in Lac Du Bois:
ldb_total_invaded_subzone <- group_by(ldb_all_sp_invaded_subzone, ZONE, SUBZONE) %>%
  summarise(total_n = n(), do_union = TRUE, .groups = "drop") %>% 
  merge(ldb_bec_ng, by = c("ZONE", "SUBZONE")) %>% 
  dplyr::mutate(total_inv_area = units::set_units(st_area(.), km^2),
                total_inv_ratio = units::drop_units(total_inv_area / ldb_bgc_area * 100))

# Total areas of invasion in Lac Du Bois
ldb_total_invaded_area <- sum(ldb_total_invaded_subzone$total_inv_area)
ldb_total_invaded_perc <- units::drop_units(ldb_total_invaded_area / ldb_bgc_area * 100)

# Summary for each invasive species in Lac Du Bois
ldb_all_sp_by_subzone <- sapply(ldb_iapp_plants, function(x) {
  dplyr::filter(ldb_all_sp_invaded_subzone, grepl(x, MAP_LABEL)) %>% 
    dplyr::mutate(n_orig = length(unique(id))) %>% 
    group_by(ZONE, SUBZONE, n_orig) %>% 
    summarise(n = n(), do_union = TRUE, .groups = "drop") %>% 
    merge(ldb_bec_ng, by = c("ZONE", "SUBZONE")) %>% 
    dplyr::mutate(inv_area = units::set_units(st_area(.), km^2),
                  inv_ratio = units::drop_units(inv_area / ldb_bgc_area * 100),
                  species = x) %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    st_cast("MULTIPOLYGON")
}, simplify = FALSE, USE.NAMES = TRUE)

# Extract spotted knapweed summary
ldb_sk_by_subzone <- ldb_all_sp_by_subzone$SK

# Total spotted knapweed invasion in BC
ldb_sk_invaded_area <- sum(ldb_sk_by_subzone$inv_area)
ldb_sk_invaded_perc <- units::drop_units(ldb_sk_invaded_area / ldb_bgc_area * 100)
ldb_sk_n_records <- ldb_sk_by_subzone$n_orig[1]

# Drop geometries for simpler outputs, adding ranking of invasion by subzone area
ldb_all_sp_by_subzone_ng <- lapply(ldb_all_sp_by_subzone, function(x) {
  st_drop_geometry(x) %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    rownames_to_column("inv_rank") %>% 
    dplyr::arrange(SUBZONE) %>% 
    dplyr::relocate(inv_rank, .after = last_col())})

# Create summary for each species by BGC Zone
ldb_all_sp_by_zone_ng <- lapply(ldb_all_sp_by_subzone_ng, function(x) {
  group_by(x, ZONE, n_orig) %>% 
    dplyr::summarise(n = sum(n, na.rm = TRUE), 
                     inv_area = sum(inv_area, na.rm = TRUE),
                     bgc_area = sum(bgc_area, na.rm = TRUE), .groups = "drop") %>% 
    dplyr::mutate(inv_ratio = units::drop_units(inv_area / bgc_area * 100),
                  bgc_ratio = units::drop_units(bgc_area / ldb_bgc_area * 100)) %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    rownames_to_column("inv_rank") %>% 
    dplyr::arrange(ZONE) %>% 
    dplyr::relocate(inv_rank, .after = last_col())})

# Finish bc_all_sp_by_subzone
ldb_all_sp_by_subzone <- lapply(ldb_all_sp_by_subzone, dplyr::select, -starts_with("bgc"))
ldb_all_sp_by_subzone_ng <- lapply(ldb_all_sp_by_subzone_ng, dplyr::select, -starts_with("bgc"))

# Extract total areas as well
ldb_all_sp_tot_area <- lapply(ldb_all_sp_by_subzone_ng, function(x) {
  summarise(x, n = sum(n), inv_area = sum(inv_area), species = unique(species)) %>% 
    dplyr::mutate(inv_ratio = units::drop_units(inv_area / ldb_bgc_area * 100))}) %>% 
  do.call(rbind, .) %>% 
  dplyr::relocate(species)

# Reorganize ldb_all_sp_by_subzone to look at invasive species in each subzone
ldb_subzone_by_sp <- sapply(ldb_bec$SUBZONE, function(x) {
  lapply(ldb_all_sp_by_subzone, dplyr::filter, SUBZONE == x) %>% 
    {if(all(sapply(., nrow) == 0)) NULL else do.call(rbind, .)}
}, simplify = FALSE, USE.NAMES = TRUE) %>% compact()

# Quick facts per subzone
ldb_top_inv_sp_by_subzone <- do.call(rbind, ldb_all_sp_by_subzone_ng) %>% 
  group_by(SUBZONE) %>% 
  slice_max(inv_ratio, n = 2)

# Again, spotted knapweed and diffuse knapweed are the top invaders in LDB when
# divided by subzone

# Create data frames for export
ldb_stats <- tribble(
  ~stat, ~value, 
  "n_invasive_recordings", nrow(ldb_iapp), 
  "n_sk_recordings", ldb_sk_n_records,
  "ldb_sk_area_km2", units::drop_units(ldb_sk_invaded_area),
  "ldb_sk_perc", ldb_sk_invaded_perc,
  "ldb_all_sp_area_km2", units::drop_units(ldb_total_invaded_area), 
  "ldb_all_sp_perc", ldb_total_invaded_perc,
  "ldb_area", units::drop_units(ldb_area)
)

# Write outputs
st_write(ldb_iapp, file.path(ldb_maps_dir, "iapp.gpkg"), delete_layer = TRUE, quiet = TRUE)
st_write(ldb_sk_by_subzone, file.path(ldb_maps_dir, "iapp_sk_by_subzone.gpkg"),
         delete_layer = TRUE, quiet = TRUE)

write.csv(ldb_stats, file.path(results_dir, "ldb_stats.csv"), row.names = FALSE)
write.csv(ldb_all_sp_tot_area, file.path(results_dir, "ldb_all_sp_areas.csv"), 
          row.names = FALSE)
write.csv(ldb_top_inv_sp_by_subzone, file.path(results_dir, "ldb_top_two_inv_by_subzone.csv"), 
          row.names = FALSE)

```

Spotted knapweed appears to be most dominant in BG zones in Lac Du Bois, followed by IDF zones. What about throughout BC? Is this consistent with the rest of the province or an artifact of a small sample area?

The below chunk takes a very long time to process (~12 - 13 hours) so prepare to leave this for a while and let it run. The reason is because the provincial BEC layer has 210 subzones and the province wide IAPP layer has >153,000 rows of data (i.e.: polygons). The intersection process that occurs thus requires 153,000 * 210 = 32,130,000 operations to run. This will only get slower as the number of IAPP polygons increase in the province.

```{r Provincial BEC and IAPP Analyses}

# This section takes a while because it deals with province wide data.
bc_boundary <- bc_bound_hres() %>% 
  dplyr::select(geometry) %>% 
  sf::st_set_agr("constant")

# Create BEC subzone layer for the entire province (~10-15 mins)
bc_bgc_subzones <- bec() %>% 
  dplyr::group_by(MAP_LABEL) %>% 
  dplyr::summarise(do_union = TRUE, .groups = "drop") %>% 
  dplyr::rename(SUBZONE = MAP_LABEL) %>% 
  sf::st_set_agr("constant") %>%
  sf::st_intersection(bc_boundary) %>% 
  dplyr::mutate(bgc_area = units::set_units(st_area(.), km^2)) %>% 
  sf::st_cast("MULTIPOLYGON") %>% 
  sf::st_set_agr("constant")

# Get total area of BC using the intersected subzone area - this is important 
# since it removes the adjoining oceanic area from mainland to island included
# in the BGC subzone layer.
bc_bgc_area <- sum(bc_bgc_subzones$bgc_area)

# Create dataframe for simpler viewing
bc_bgc_subzones_ng <- st_drop_geometry(bc_bgc_subzones) %>% 
  dplyr::mutate(bgc_ratio = units::drop_units(bgc_area / bc_bgc_area * 100)) %>% 
  arrange(desc(bgc_ratio)) %>% 
  rownames_to_column("size_rank") %>% 
  arrange(SUBZONE) %>% 
  dplyr::relocate(size_rank, .after = last_col())

# Get full IAPP database, filtering the pure "ZZ" calls out (~10-15 minutes)
bc_iapp <- bcdata::bcdc_query_geodata("10ecf9ad-1555-4043-834a-f5d24a506d59", crs = 3005) %>% 
  collect() %>% 
  dplyr::filter(MAP_LABEL != "ZZ") %>% 
  st_cast("MULTIPOLYGON") %>% 
  sf::st_set_agr("constant")

# Extract vector of plant codes that occur throughout the entire IAPP
bc_iapp_plants <- sapply(iapp_plants, grep, bc_iapp$MAP_LABEL) %>% 
  compact() %>% names()

# This part takes a very long time, ~ 12 hours to do full IAPP intersection with 
# BGC subzone layer
bc_all_sp_invaded_subzone <- sf::st_intersection(bc_iapp, bc_bgc_subzones)

# Summary of the full IAPP dataset in BC by BGC subzone:
bc_total_invaded_subzone <- group_by(bc_all_sp_invaded_subzone, SUBZONE) %>%
  summarise(total_n = n(), do_union = TRUE, .groups = "drop") %>% 
  merge(bc_bgc_subzones_ng, by = "SUBZONE") %>% 
  dplyr::mutate(total_inv_area = units::set_units(st_area(.), km^2),
                total_inv_ratio = units::drop_units(total_inv_area / bgc_area * 100))

# Total areas of invasion in BC
bc_invaded_area <- sum(bc_total_invaded_subzone$total_inv_area)
bc_invaded_perc <- units::drop_units(bc_invaded_area / bc_bgc_area * 100)

# Summary for each invasive species in BC:
bc_all_sp_by_subzone <- sapply(bc_iapp_plants, function(x) {
  dplyr::filter(bc_all_sp_invaded_subzone, grepl(x, MAP_LABEL)) %>% 
    dplyr::mutate(n_orig = length(unique(id))) %>% 
    group_by(SUBZONE, n_orig) %>% 
    summarise(n = n(), do_union = TRUE, .groups = "drop") %>% 
    merge(bc_bgc_subzones_ng, by = "SUBZONE") %>% 
    dplyr::mutate(inv_area = units::set_units(st_area(.), km^2),
                  inv_ratio = units::drop_units(inv_area / bgc_area * 100)) %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    st_cast("MULTIPOLYGON")
}, simplify = FALSE, USE.NAMES = TRUE)

# Refactor to look at invasion of each species by subzone
bc_inv_subzones <- sort(unique(unlist(
  sapply(bc_all_sp_by_subzone, dplyr::pull, SUBZONE))))
bc_all_subzones_by_sp <- sapply(bc_inv_subzones, function(x) {
  do.call(rbind, lapply(bc_all_sp_by_subzone, dplyr::filter, SUBZONE == x)) %>% 
    rownames_to_column("species") %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    dplyr::mutate(n_total = sum(n), )
}, simplify = FALSE, USE.NAMES = TRUE)

# Extract the spotted knapweed subzone summary for BC
bc_sk_by_subzone <- bc_all_sp_by_subzone$SK

# Total spotted knapweed invasion in BC
bc_sk_invaded_area <- sum(bc_sk_by_subzone$inv_area)
bc_sk_invaded_perc <- units::drop_units(bc_sk_invaded_area / bc_bgc_area * 100) 
bc_sk_n_records <- bc_sk_by_subzone$n_orig[1]

# Remove geometry columns from total and spotted knapweed only to perform other
# relevant calculations for output - "ng" = "no geometry"
# First, do the bc_total_invaded_subzone to get whole subzone summaries
bc_total_invaded_subzone_ng <- st_drop_geometry(bc_total_invaded_subzone) %>% 
  dplyr::arrange(desc(total_inv_ratio)) %>% 
  rownames_to_column("total_inv_rank") %>% 
  dplyr::arrange(SUBZONE) %>% 
  dplyr::relocate(total_inv_rank, .after = last_col()) %>% 
  dplyr::select(-starts_with("bgc"))

# Next, drop geometry for summaries by species:
bc_all_sp_by_subzone_ng <- lapply(bc_all_sp_by_subzone, function(x) {
  st_drop_geometry(x) %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    rownames_to_column("inv_rank") %>% 
    dplyr::arrange(SUBZONE) %>% 
    dplyr::relocate(inv_rank, .after = last_col())})

# Extract the spotted knapweed summary
bc_sk_by_subzone_ng <- bc_all_sp_by_subzone_ng$SK %>% 
  dplyr::select(-starts_with("bgc"))

# Recreate a "ZONE" column and create summaries of total invaded area by BGC zone
bc_total_invaded_zone_ng <- bc_total_invaded_subzone_ng %>% 
  merge(merge(bc_sk_by_subzone_ng, bc_bgc_subzones_ng, all = TRUE), all = TRUE) %>% 
  dplyr::mutate(ZONE = sapply(strsplit(SUBZONE, "[[:lower:]]"), "[[", 1)) %>% 
  group_by(ZONE) %>% 
  dplyr::summarise(total_n = sum(total_n, na.rm = TRUE), 
                   total_inv_area = sum(total_inv_area, na.rm = TRUE),
                   bgc_area = sum(bgc_area, na.rm = TRUE),
                   sk_n_orig = mean(n_orig, na.rm = TRUE), 
                   sk_n = sum(n, na.rm = TRUE),
                   sk_inv_area = sum(inv_area, na.rm = TRUE), .groups = "drop") %>% 
  dplyr::mutate(total_inv_ratio = units::drop_units(total_inv_area / bgc_area * 100),
                bgc_ratio = units::drop_units(bgc_area / bc_bgc_area * 100),
                sk_ratio = units::drop_units(sk_inv_area / bgc_area * 100))

# Recreate a "ZONE" column and create summaries by BGC zone for each species
bc_all_sp_by_zone_ng <- lapply(bc_all_sp_by_subzone_ng, function(x) {
  dplyr::mutate(x, ZONE = sapply(strsplit(SUBZONE, "[[:lower:]]"), "[[", 1)) %>% 
    group_by(ZONE) %>% 
    dplyr::summarise(n = sum(n, na.rm = TRUE), 
                     inv_area = sum(inv_area, na.rm = TRUE),
                     bgc_area = sum(bgc_area, na.rm = TRUE),
                     n_orig = mean(n_orig, na.rm = TRUE),  .groups = "drop") %>% 
    dplyr::mutate(inv_ratio = units::drop_units(inv_area / bgc_area * 100),
                  bgc_ratio = units::drop_units(bgc_area / bc_bgc_area * 100)) %>% 
    dplyr::arrange(desc(inv_ratio)) %>% 
    rownames_to_column("inv_rank") %>% 
    dplyr::arrange(ZONE) %>% 
    dplyr::relocate(inv_rank, .after = last_col())})

# Finish bc_all_sp_by_subzone
bc_all_sp_by_subzone <- lapply(bc_all_sp_by_subzone, dplyr::select, -starts_with("bgc"))
bc_all_sp_by_subzone_ng <- lapply(bc_all_sp_by_subzone_ng, dplyr::select, -starts_with("bgc"))

# Extract total areas as well
bc_all_sp_tot_area <- lapply(bc_all_sp_by_subzone_ng, function(x) {
  summarise(x, n = mean(n_orig, na.rm = TRUE), inv_area = sum(inv_area)) %>% 
    dplyr::mutate(inv_ratio = units::drop_units(inv_area / bc_bgc_area * 100))}) %>% 
  do.call(rbind, .) %>% 
  rownames_to_column("species")
    
# Create tables for export: 
# Export BEC data by subzone
bc_bgc_data <- merge(bc_bgc_subzones_ng, merge(
  bc_total_invaded_subzone_ng, bc_sk_by_subzone_ng, all = TRUE), all = TRUE) %>% 
  dplyr::select(SUBZONE, starts_with("bgc"), ends_with("rank")) %>% 
  dplyr::rename(sk_inv_rank = inv_rank)

# Export invasion data by subzone - get most invasive species by subzone first
bc_most_inv_sp_by_subzone <- lapply(bc_all_subzones_by_sp, dplyr::slice_max, inv_ratio) %>% 
  do.call(rbind, .) %>% 
  st_drop_geometry() %>% 
  group_by(SUBZONE) %>% 
  dplyr::slice_max(inv_area) %>% 
  dplyr::slice_max(n) %>% 
  dplyr::slice_max(n_orig) %>% 
  ungroup() %>% 
  dplyr::select(species, SUBZONE)

bc_inv_data <- Reduce(
  function(x, y, ...) merge(x, y, all = TRUE, ...),
  list(bc_bgc_subzones_ng, bc_total_invaded_subzone_ng, bc_sk_by_subzone_ng, 
       bc_most_inv_sp_by_subzone)) %>% 
  dplyr::select(-c(starts_with("bgc"), ends_with("rank"))) %>% 
  dplyr::rename(sk_n = n, sk_n_orig = n_orig, sk_inv_ratio = inv_ratio, 
                sk_inv_area = inv_area, most_inv_species = species)

# Filter to only include subzone names that are in LDB for easier comparison
bc_bgc_data_limit <- dplyr::filter(bc_bgc_data, SUBZONE %in% ldb_bec$SUBZONE)
bc_inv_data_limit <- dplyr::filter(bc_inv_data, SUBZONE %in% ldb_bec$SUBZONE)

bc_extra_data <- tribble(
  ~stat, ~value, 
  "total_inv_area_km2", units::drop_units(bc_invaded_area),
  "total_inv_ratio", bc_invaded_perc,
  "total_inv_records", nrow(bc_iapp),
  "sk_inv_area_km2", units::drop_units(bc_sk_invaded_area),
  "sk_inv_ratio", bc_sk_invaded_perc,
  "sk_inv_records", bc_sk_n_records,
  "total_bc_area_km2", units::drop_units(bc_bgc_area))

# Summary by species
bc_inv_sp_extent <- data.frame(n_subzones = sapply(bc_all_sp_by_subzone, nrow),
                               inv_area = sapply(bc_all_sp_by_subzone, function(x)
                                 sum(dplyr::pull(x, inv_area))),
                               n_records = sapply(bc_all_sp_by_subzone, function(x)
                                 dplyr::pull(x, n_orig)[1]),
                               most_inv_subzone = sapply(bc_all_sp_by_subzone, function(x)
                                 dplyr::slice_max(x, inv_area)$SUBZONE)) %>% 
  rownames_to_column("species") %>% 
  dplyr::arrange(desc(inv_area))

# Save as individual files for faster reading in later on if necessary
st_write(bc_bgc_subzones, file.path(bc_maps_dir, "bc_subzones.gpkg"), delete_layer = TRUE, quiet = TRUE)
st_write(bc_iapp, file.path(bc_maps_dir, "bc_iapp.gpkg"), delete_layer = TRUE, quiet = TRUE)
st_write(bc_sk_by_subzone, file.path(bc_maps_dir, "bc_sk_by_subzone.gpkg"), delete_layer = TRUE, quiet = TRUE)
st_write(bc_all_sp_invaded_subzone, file.path(bc_maps_dir, "bc_all_sp_invaded_subzone.gpkg"), 
         delete_layer = TRUE, quiet = TRUE)

write.csv(bc_bgc_data, file.path(results_dir, "bc_bgc_data_total.csv"), row.names = FALSE)
write.csv(bc_bgc_data_limit, file.path(results_dir, "bc_bgc_data_limit.csv"), row.names = FALSE)
write.csv(bc_inv_data, file.path(results_dir, "bc_invasive_data_total.csv"), row.names = FALSE)
write.csv(bc_inv_data_limit, file.path(results_dir, "bc_invasive_data_limit.csv"), row.names = FALSE)
write.csv(bc_extra_data, file.path(results_dir, "bc_extra_data.csv"), row.names = FALSE)
write.csv(bc_total_invaded_zone_ng, file.path(results_dir, "bc_zone_inv_data.csv"), row.names = FALSE)
write.csv(bc_inv_sp_extent, file.path(results_dir, "bc_inv_sp_extent.csv"), row.names = FALSE)

```

Some takeaways from the analysis above: 
SK covers ~.056% of the province of BC. That's not overly concerning; however, it does take up ~3% of the BG zone (7891 ha) and 2.46% of the PP zone (7260 ha), with the largest by area invasion happenning in the IDF zone (31,6677 ha), though this zone is very large and only accounts for 0.71% of the entire invaded area. This is important: this implies that the rate of spread in Lac Du Bois park (24%, 1.8%, and 9.1% in respective zones) is generally much higher than the remainder of the province for some reason!

Each zone has very important values for forestry, First Nations, and conservation of endangered species. With that in mind, this process for finding invasive species trends throughout our province should be looked at more seriously. There are currently nearly 160,000 polygons for invasive plant records throughout BC, perhaps we should see where SK falls in terms of how invasive it is in comparison to other plant species. This will require a further investigation into the text of the "MAP_LABEL" column, below:

The analysis above shows that SK and DK are the most invasive plants in BC. Additionally, even though only 0.05% of the province is covered in SK, all invasive plants cover only 0.145% of the province, thus SK accounts for slightly more than 1/3 of the invasive plant cover in BC.

```{r BC Raster analysis}

dl_dir <- file.path(tempdir()) # Change to suit your needs

# URL's to download tiles for Lac Du Bois
urls <- c("https://pub.data.gov.bc.ca/datasets/175624/92i/092i15_e.dem.zip", 
          "https://pub.data.gov.bc.ca/datasets/175624/92i/092i16_w.dem.zip",
          "https://pub.data.gov.bc.ca/datasets/175624/92i/092i10_e.dem.zip", 
          "https://pub.data.gov.bc.ca/datasets/175624/92i/092i09_w.dem.zip")

# Merge the files into a single object
trim <- lapply(urls, function(i) {
  download.file(url = i, destfile = file.path(dl_dir, basename(i)))
  z <- unzip(file.path(dl_dir, basename(i)), overwrite = TRUE, 
             exdir = file.path(dl_dir))
  unlink(file.path(dl_dir, basename(i)))
  return(rast(z))
}) %>% do.call(merge, .) %>% magrittr::set_names("dem")

# Convert to raster object, crop to shape of lac du bois with 1 cell (~25m) buffer
trim_rast <- raster(trim)
trim_rast <- crop(trim_rast, extent(st_transform(st_buffer(ldb_bec, 25), 4326)))
trim_rast[trim_rast < 0] <- -9999
NAvalue(trim_rast) <- -9999

# Save this DEM as a .asc file to be used in ClimateBC program
writeRaster(trim_rast, "./trim.asc", overwrite = TRUE)

# Do proper raster datum transformation through raster package, then read as 
# terra package object when complete
trim_rast_ext <- projectExtent(trim_rast, crs("+init=epsg:3005"))
res(trim_rast_ext) <- max(res(trim_rast_ext))
trim_ext <- rast(trim_rast_ext)
crs(trim_ext) <- "epsg:3005"

# Do reprojection with now proper extent, and mask to LDB subzone layer
trim_bca <- project(trim, trim_ext)
trim_mask <- mask(trim_bca, vect(ldb_bec)) %>% 
  terra::trim()

# Create BGC subzone dataframe to merge data with later on
ldb_bec2 <- rownames_to_column(ldb_bec, "ID") %>% 
  st_drop_geometry() %>% 
  dplyr::select(-bgc_area)

# Extract mean, min, and max elevations by subzone in LDB
trim_extract <- as.data.frame(extract(trim_mask, vect(ldb_bec))) %>% 
  merge(ldb_bec2, by = "ID") %>% 
  group_by(SUBZONE) %>% 
  summarise(min = min(dem), max = max(dem), mean = mean(dem), .groups = "drop")

# Load temperature mean, min, and max rasters. Divide by 10 to get real values
# of temperature instead of integers
min_sum <- rast("./trim/Year_2017S/Tmin_sm.asc") / 10
max_sum <- rast("./trim/Year_2017S/Tmax_sm.asc") / 10
avg_sum <- rast("./trim/Year_2017S/Tave_sm.asc") / 10

# Get min, max, and mean temperatures for each subzone in LDB by reprojecting,
# masking, and extracting the temperature rasters
min_subzone <- project(min_sum, "epsg:3005") %>% 
  mask(vect(ldb_bec)) %>% 
  extract(vect(ldb_bec)) %>% 
  as.data.frame() %>% 
  merge(ldb_bec2, by = "ID") %>% 
  group_by(SUBZONE) %>% 
  summarise(min = min(Tmin_sm), .groups = "drop")

max_subzone <- project(max_sum, "epsg:3005") %>% 
  mask(vect(ldb_bec)) %>% 
  extract(vect(ldb_bec)) %>% 
  as.data.frame() %>% 
  merge(ldb_bec2, by = "ID") %>% 
  group_by(SUBZONE) %>% 
  summarise(max = max(Tmax_sm), .groups = "drop")

avg_subzone <- project(avg_sum, "epsg:3005") %>% 
  mask(vect(ldb_bec)) %>% 
  extract(vect(ldb_bec)) %>% 
  as.data.frame() %>% 
  merge(ldb_bec2, by = "ID") %>% 
  group_by(SUBZONE) %>% 
  summarise(mean = mean(Tave_sm), .groups = "drop")

subzone_clim <- merge(min_subzone, merge(max_subzone, avg_subzone))
write.csv(trim_extract, file.path(results_dir, "elev_by_subzone.csv"), row.names = FALSE)
write.csv(subzone_clim, file.path(results_dir, "temp_by_subzone.csv"), row.names = FALSE)

```

Next, I want to pull in the spatial data gathered from field work and see what I can capture from that:

```{r MSC Study area analysis}

msc_map_dir <- file.path("./01_Spatial_Data/layers/study_area")
dir.create(msc_map_dir, showWarnings = FALSE, recursive = TRUE)

unzip(file.path(data_dir, "Matt.kmz"), exdir = tempdir(), overwrite = TRUE)
layers <- st_layers(file.path(tempdir(), "doc.kml"))$name
field_data <- sapply(layers, function(x)
  st_read(file.path(tempdir(), "doc.kml"), layer = x, quiet = TRUE), 
  simplify = FALSE, USE.NAMES = TRUE)
area_of_invasion <- field_data[[which(sapply(field_data, function(x) 
  st_geometry_type(x, by_geometry = FALSE) == "POLYGON"))]] %>% 
  st_transform(3005) %>% 
  st_make_valid()
sample_sites <- field_data[which(sapply(field_data, function(x) 
  st_geometry_type(x, by_geometry = FALSE) == "POINT" && any(grepl("Pri|Inv", x$Name))))] %>% 
  do.call(rbind, .) %>% 
  st_transform(3005) %>% 
  st_make_valid()

# Do some exploration of the spatial data here:
sample_site_exp <- cbind(
  sample_sites, 
  within = st_within(sample_sites, area_of_invasion, sparse = FALSE), 
  distance_to_poly = st_distance(sample_sites, area_of_invasion))
aoi_within <- st_union(area_of_invasion) %>% 
  st_intersection(ldb_sk)
aoi_exp <- area_of_invasion %>% 
  dplyr::mutate(area = st_area(.)) %>% 
  dplyr::mutate(within_iapp_poly = units::drop_units(st_area(aoi_within) / area)) %>% 
  dplyr::mutate(beyond_iapp_poly = 1 - within_iapp_poly) %>% 
  st_drop_geometry() %>% 
  dplyr::select(-Description)

# Write layers separately
st_write(area_of_invasion, file.path(msc_map_dir, "aoi.gpkg"), delete_layer = TRUE,
         quiet = TRUE)
st_write(sample_site_exp, file.path(msc_map_dir, "sample_locations.gpkg"), 
         delete_layer = TRUE, quiet = TRUE)
write.csv(aoi_exp, file.path(results_dir, "study_area_info.csv"), row.names = FALSE)
st_write(sample_sites, file.path(results_dir, "study_site_coordinates.csv"), 
         layer_options = "GEOMETRY=AS_XY", delete_dsn = TRUE)

```

I have all of the necessary shapefiles that I would need. Now, I'd like to create a visually appealing figure for use in my thesis. I tried for a while to do this in R but found it way too complicated, so I moved to QGIS for more visually appealing figures.

Separately, here is some code for finding all range reference areas within a study area (may also include smaller fencing units as well that are used in rangelands). It took a bit of wrangling but this works decently well, I think...

```{r grazing exclosures}
# AKA range reference areas (RRA's)

ldb_rra <- bcdata::bcdc_query_geodata("946b09a4-c63e-48eb-b322-3701c8ba161d", crs = 3005) %>% 
  bcdata::filter(bcdata::BBOX(sf::st_bbox(ldb_aoi), crs = 3005)) %>% 
  collect() %>% 
  sf::st_set_agr("constant") %>% 
  {if(nrow(.) > 0) {
    sf::st_intersection(., ldb_aoi)
  } else .}

# First, fix geometries. Some lines are close to other lines which should be merged
# together. Doing a merge creates duplicate geometries which will be filtered out later
ldb_rra_geom <- st_geometry(ldb_rra)
ldb_rra_geom <- lapply(1:length(ldb_rra_geom), function(xx) {
  x <- ldb_rra_geom[[xx]]
  if(st_geometry_type(x) != "LINESTRING") NULL else {
    ind <- which(st_is_within_distance(st_sfc(x, crs = 3005), ldb_rra_geom, dist = 1, sparse = FALSE))
    ind <- ind[ind != xx]
    ind <- ind[st_geometry_type(ldb_rra_geom[ind]) == "LINESTRING"]
    if(length(ind) > 0) {
      y <- st_union(
        st_sfc(x, crs = 3005), 
        ldb_rra_geom[ind]
      )
      coords <- st_coordinates(y)[, c("X", "Y")]
      if(nrow(coords) >= 5) {
        z <- foreach(i = 1:length(y), .combine = rbind, .final = st_linestring) %do% {
          st_coordinates(y[[i]])[, c("X", "Y")]
        } %>% st_cast("POLYGON")
        if(st_is_valid(z)) z else {
          zz <- st_make_valid(z)
          if(st_geometry_type(zz) == "GEOMETRYCOLLECTION") {
            polys <- st_collection_extract(zz, "POLYGON")
            lines <- st_collection_extract(zz, "LINESTRING")
            if((length(polys) != 1 || st_geometry_type(polys) != "POLYGON") ||
               st_geometry_type(lines) != "LINESTRING") {
              NULL
            } else {
              polys
            }
          } else NULL
        }
      } else NULL
    } else {
      coords <- st_coordinates(x)[, c("X", "Y")]
      if(nrow(coords) >= 5) {
        p1 <- st_point(coords[1, ])
        p2 <- st_point(coords[nrow(coords), ])
        if(st_is_within_distance(p1, p2, 0, sparse = FALSE)) {
          st_cast(x, "POLYGON")
        } else NULL
      } else NULL
    }
  }
})

# Apply new geometries to geometry columns
ldb_rra <- st_set_geometry(ldb_rra, st_sfc(ldb_rra_geom, crs = 3005)) %>% 
  dplyr::filter(!st_is_empty(.))

# Remove duplicated polygons
ldb_rra_geom <- st_geometry(ldb_rra)
ldb_rra_geom <- lapply(1:length(ldb_rra_geom), function(xx) {
  x <- ldb_rra_geom[[xx]]
  sub_geom <- lapply(1:length(ldb_rra_geom), function(yy) {
    if(yy > xx) {
      y <- ldb_rra_geom[[yy]]
      if(x == y) {
        out <- NULL 
      } else out <- x
    } else out <- x
  })
  if(any(sapply(sub_geom, is.null))) NULL else x
})

# Apply new geometries to geometry columns
ldb_rra <- st_set_geometry(ldb_rra, st_sfc(ldb_rra_geom, crs = 3005)) %>% 
  dplyr::filter(!st_is_empty(.))

# Do a size ratio filter, keep the fence for the McQueen Lake ecological reserve, 
# and remove a single offending polygon
area_perim <- data.frame(
  rows = as.numeric(rownames(ldb_rra)),
  id = ldb_rra$id,
  area = st_area(ldb_rra), 
  perim = ldb_rra %>% st_cast("MULTILINESTRING") %>% st_length()
) %>% dplyr::mutate(ratio = as.numeric(area / perim)) %>% 
  dplyr::mutate(ratio = ifelse(grepl("380b|654d", id), 1, ratio), 
                area = ifelse(grepl("380b|654d", id), 1, area)) %>% 
  dplyr::filter(ratio >= 1, ratio < 20, area < 10000, !endsWith(id, c("377b", "2fb", "74ca", "65dd")))

# Filter ldb_rra dataframe using matching rownames from area/perimeter filterings
ldb_rra <- ldb_rra[area_perim$rows, ]

# Write output
st_write(ldb_rra, file.path(ldb_maps_dir, "rras.gpkg"), delete_layer = TRUE, quiet = TRUE)

```
