==> 02_Greenhouse_Trials.Rmd <==
---
title: "02_Greenhouse_Trials"
author: "Matt"
date: "9/14/2020"
output: html_document
---

The intention of this document is to perform scientific analysis of the greenhouse trials performed in my MSc. This will include everything from greenhouse temperature data to biomass and carbon sampling as well.

I'll start with biomass collected in the greenhouse. First, load some packages

```{r Load packages}

suppressMessages(suppressWarnings({
  ls <- c("tidyverse", "broom", "car", "ggpubr", "magick", "agricolae", 
          "extrafont", "rstatix", "emmeans", "ARTool")
  new_packages <- ls[!(ls %in% installed.packages()[, "Package"])]
  if(length(new_packages)) install.packages(new_packages)
  sapply(ls, library, character.only = TRUE, quietly = TRUE)[0]
  if(any(!fonts() %in% unlist(windowsFonts())) || length(fonts()) == 0) {
    font_import(prompt = FALSE)
    loadfonts(device = "win")}
  rm(ls, new_packages)}))

```

Next, initialize folders and load the data

```{r Set directories and load biomass data}

data_dir <- file.path("./02_Greenhouse/data")
plot_dir <- file.path("./02_Greenhouse/plots")
results_dir <- file.path("./02_Greenhouse/results")
dir.create(plot_dir, showWarnings = FALSE)
dir.create(results_dir, showWarnings = FALSE)

biomass <- read.csv(file.path(data_dir, "greenhouse_raw.csv"), header = TRUE) %>% 
  dplyr::mutate(Soil_Type = factor(Soil_Type, levels = c("Invaded", "Pristine")),
                Soil_Treatment = factor(Soil_Treatment, levels = c("Con", "AC", "Ash")), 
                AboveMass = ifelse(AboveMass == 0.001, 0, AboveMass),
                BelowMass = ifelse(BelowMass == 0.001, 0, BelowMass),
                TotalMass = AboveMass + BelowMass) %>% 
  dplyr::select(Soil_Type, Soil_Treatment, Species, TotalMass)

```

Let's see if the data are normal or not. This requires separating by species since they have very different growth patterns etc.

For both species, test if residuals are significantly different (lots of variation). If so, the data are not normal and something needs to be done to it.

```{r biomass stats}

anova_assum <- function(x, dv, rm_outliers = TRUE) {
  sapply(unique(x$Species), function(i) {
    df <- dplyr::filter(x, Species == i) %>% 
      dplyr::rename(var = sym(dv)) %>% 
      dplyr::select(Soil_Type, Soil_Treatment, Species, var) %>% 
      group_by(Soil_Type, Soil_Treatment) %>% 
      mutate(outlier = is_outlier(var), extreme = is_extreme(var)) %>% 
      ungroup() %>% 
      {if(rm_outliers) dplyr::filter(., !extreme) else .}
    norm_model <- lm(var ~ Soil_Type * Soil_Treatment, data = df)
    norm_plot_model <- ggqqplot(residuals(norm_model))
    shapiro_model <- shapiro_test(residuals(norm_model))
    groups <- df %>%
      group_by(Soil_Type, Soil_Treatment) %>% 
      dplyr::summarise(., n = n_distinct(var), .groups = "drop") %>% 
      pull(n)
    shapiro_groups <- if(any(groups == 1)) {
      data.frame(Soil_Type = NA, Soil_Treatment = NA, variable = dv, statistic = NA,
                 p = "Could not calculate Shapiro - only one unique value for one or some groups")
    } else {
      df %>%
        group_by(Soil_Type, Soil_Treatment) %>% 
        shapiro_test(var) %>% 
        dplyr::mutate(variable = dv)
    }
    norm_plot_groups <- ggqqplot(df, "var", ggtheme = theme_bw()) + 
      facet_grid(Soil_Type ~ Soil_Treatment)
    levene <- df %>% levene_test(var ~ Soil_Type * Soil_Treatment)
    df <- dplyr::rename(df, !!dv := var)
    return(list(data = df, shapiro_whole = shapiro_model, shapiro_groups = shapiro_groups,
                norm_plot_whole = norm_plot_model, norm_plot_groups = norm_plot_groups, 
                levene = levene))
  }, simplify = FALSE, USE.NAMES = TRUE)
}

bmass_assum <- anova_assum(biomass, "TotalMass", rm_outliers = FALSE)

bmass_normal_analysis <- lapply(bmass_assum, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
bmass_var_eq_analysis <- lapply(bmass_assum, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

# Base data: Neither RF or SK pass normality assumptions. Only SK passes 
# equal variance assumption. Keep in mind, this is without filtering away the 
# outlier data yet. 

# Square root transforming will work with either the unfiltered or the filtered
# data. It will depend on the committee: If no extreme outliers is a consideration, 
# I will have to use this data transformation instead
bmass_sqrt <- dplyr::mutate(biomass, sqrt_mass = sqrt(TotalMass))
bmass_assum <- anova_assum(bmass_sqrt, "sqrt_mass", rm_outliers = FALSE)
bmass_normal_analysis <- lapply(bmass_assum, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
bmass_var_eq_analysis <- lapply(bmass_assum, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

# The results came back passing all assumptions, and now a proper 2-way ANOVA can
# be carried out (and post hoc analysis, Tukey HSD):

anova_analysis <- function(x, dv) {
  sapply(unique(x$Species), function(i) {
    df <- dplyr::filter(x, Species == i) %>% 
      dplyr::rename(var = sym(dv)) %>% 
      droplevels() %>% 
      dplyr::mutate(Soil_Type = factor(Soil_Type, levels = c("Invaded", "Pristine")),
                    Soil_Treatment = factor(Soil_Treatment, levels = c("Con", "Ash", "AC"))) %>% 
      as.data.frame()
    fct <- ifelse(dv %in% c("sqrt_mass", "Carbon"), 3, ifelse(dv == "Nitrogen", 0.5, 1))
    aov <- df %>% anova_test(var ~ Soil_Type * Soil_Treatment)
    model <- lm(var ~ Soil_Type * Soil_Treatment, data = df)
    aov_type <- df %>% 
      group_by(Soil_Treatment) %>% 
      anova_test(var ~ Soil_Type, error = model)
    aov_treatment <- df %>% 
      group_by(Soil_Type) %>% 
      anova_test(var ~ Soil_Treatment, error = model)
    
    pwc_type <- group_by(df, Soil_Type) %>% 
      {if(all(aov_type$p < 0.05)) {
        emmeans_test(., var ~ Soil_Treatment, p.adjust.method = "bonferroni") 
    } else {
        pairwise_t_test(., var ~ Soil_Treatment, p.adjust.method = "bonferroni")
    }} %>% 
      add_xy_position(x = "Soil_Type", fun = "mean_sd", step.increase = 0) %>% 
      dplyr::mutate(y.position = y.position + 0.1 * fct, term = "Soil_Treatment",
                    p_round = round(p.adj, digits = 3)) %>% 
      dplyr::select(-any_of(c("groups", "n1", "n2", "p.signif", "df"))) %>% 
      as.data.frame()
    
    # pairwise_t_test and pairwise.t.test functions are evil. Corrections made here.
    if(!all(aov_type$p < 0.05)) {
      pwc_type <- lapply(unique(df$Soil_Type), function(xx) {
        df <- df %>% dplyr::filter(Soil_Type == xx)
        xbar <- tapply(df$var, df$Soil_Treatment, mean, na.rm = TRUE)
        s <- tapply(df$var, df$Soil_Treatment, sd, na.rm = TRUE)
        n <- tapply(!is.na(df$var), df$Soil_Treatment, sum)
        degf <- n - 1
        total.degf <- sum(degf)
        pooled.sd <- sqrt(sum(s^2 * degf)/total.degf)
        compare.levels <- function(i, j) {
          dif <- xbar[i] - xbar[j]
          se.dif <- pooled.sd * sqrt(1/n[i] + 1/n[j])
          t.val <- dif/se.dif
          return(list(t = t.val, p = 2 * pt(-abs(t.val), total.degf)))
        }
        ix <- setNames(seq_along(levels(df$Soil_Treatment)), levels(df$Soil_Treatment))
        tt <- outer(ix[-1L], ix[-length(ix)], function(ivec, jvec) 
          vapply(seq_along(ivec), 
                 function(k) {
                   i <- ivec[k]
                   j <- jvec[k]
                   if (i > j) 
                     compare.levels(i, j)$t
                   else NA_real_
                 }, 0.05))
        
        tt_df <- as.data.frame(tt) %>% 
          rownames_to_column("group2") %>% 
          pivot_longer(2:3, names_to = "group1", values_to = "statistic") %>% 
          na.omit()
        
        pwc_out <- merge(pwc_type %>% dplyr::filter(Soil_Type == xx), tt_df, 
                         by = c("group1", "group2"))
      }) %>% do.call(rbind, .)
    }
    
    pwc_type <- rbind(
      dplyr::filter(pwc_type, x == 1) %>% 
        dplyr::mutate(y.position = seq(
        from = min(.$y.position), by = 0.2 * fct, length.out = 3)),
      dplyr::filter(pwc_type, x == 2) %>% 
        dplyr::mutate(y.position = seq(
        from = min(.$y.position), by = 0.2 * fct, length.out = 3)))
    
    pwc_type <- rbind(
      dplyr::filter(pwc_type, x == 1) %>% 
        {if(dplyr::slice(., which.min(y.position)) %>% 
            dplyr::pull(p.adj) >= 0.05 && any(.$p.adj < 0.05)) {
          tmp <- .
          min_y <- min(tmp$y.position)
          n_inc <- nrow(dplyr::filter(tmp, p.adj < 0.05))
          if(n_inc == 2) {
            sq <- c(min_y, seq(min_y, by = 0.2 * fct, length.out = n_inc))
          } else {
            sq <- rep(min_y, 3)
          }
          dplyr::mutate(., y.position = sq)
        } else . } %>% 
        {if(.$p.adj[2] >= 0.05) {
          dplyr::mutate(., y.position = c(y.position[1], y.position[2], y.position[2]))
        } else . }, 
      dplyr::filter(pwc_type, x == 2) %>% 
        {if(dplyr::slice(., which.min(y.position)) %>% 
            dplyr::pull(p.adj) >= 0.05 && any(.$p.adj < 0.05)) {
          tmp <- .
          min_y <- min(tmp$y.position)
          n_inc <- nrow(dplyr::filter(tmp, p.adj < 0.05))
          if(n_inc == 2) {
            sq <- c(min_y, seq(min_y, by = 0.2 * fct, length.out = n_inc))
          } else {
            sq <- rep(min_y, 3)
          }
          dplyr::mutate(., y.position = sq)
        } else . } %>% 
        {if(.$p.adj[2] >= 0.05) {
          dplyr::mutate(., y.position = c(y.position[1], y.position[2], y.position[2]))
        } else . })
    
    pwc_treatment <- group_by(df, Soil_Treatment) %>% 
      {if(all(aov_treatment$p < 0.05)) {
        emmeans_test(., var ~ Soil_Type, p.adjust.method = "bonferroni")
    } else {
        pairwise_t_test(., var ~ Soil_Type, p.adjust.method = "bonferroni")
    }} %>% 
      add_xy_position(x = "Soil_Treatment", fun = "mean_sd", step.increase = 0) %>% 
      dplyr::mutate(y.position = y.position + 0.1 * fct, term = "Soil_Type",
                    p_round = round(p.adj, digits = 3)) %>% 
      dplyr::select(-any_of(c("groups", "n1", "n2", "p.signif", "df"))) %>% 
      as.data.frame()
    
    # pairwise_t_test and pairwise.t.test functions are evil. Corrections made here.
    if(!all(aov_treatment$p < 0.05)) {
      pwc_treatment <- lapply(unique(df$Soil_Treatment), function(xx) {
        df <- df %>% dplyr::filter(Soil_Treatment == xx)
        xbar <- tapply(df$var, df$Soil_Type, mean, na.rm = TRUE)
        s <- tapply(df$var, df$Soil_Type, sd, na.rm = TRUE)
        n <- tapply(!is.na(df$var), df$Soil_Type, sum)
        degf <- n - 1
        total.degf <- sum(degf)
        pooled.sd <- sqrt(sum(s^2 * degf)/total.degf)
        compare.levels <- function(i, j) {
          dif <- xbar[i] - xbar[j]
          se.dif <- pooled.sd * sqrt(1/n[i] + 1/n[j])
          t.val <- dif/se.dif
          return(list(t = t.val, p = 2 * pt(-abs(t.val), total.degf)))
        }
        ix <- setNames(seq_along(levels(df$Soil_Type)), levels(df$Soil_Type))
        tt <- outer(ix[-1L], ix[-length(ix)], function(ivec, jvec) 
          vapply(seq_along(ivec), 
                 function(k) {
                   i <- ivec[k]
                   j <- jvec[k]
                   if (i > j) 
                     compare.levels(i, j)$t
                   else NA_real_
                 }, 0.05))
        
        tt_df <- as.data.frame(tt) %>% 
          rownames_to_column("group2") %>% 
          pivot_longer(2, names_to = "group1", values_to = "statistic") %>% 
          na.omit()
        
        pwc_out <- merge(pwc_treatment %>% dplyr::filter(Soil_Treatment == xx), tt_df, 
                         by = c("group1", "group2"))
      }) %>% do.call(rbind, .)
    }
    
    tukey <- df %>% 
      tukey_hsd(var ~ Soil_Type * Soil_Treatment) %>% 
      adjust_pvalue(method = "bonferroni")
    
    df_sum <- df %>% 
      group_by(Soil_Type, Soil_Treatment) %>% 
      summarise(across(var, list(mean = ~mean(.x, na.rm = TRUE),
                                 sd = ~sd(.x, na.rm = TRUE)),
                       .names = "{.fn}"), .groups = "drop")
    
    df <- dplyr::rename(df, !!dv := var)
    
    # bxp_type <- ggboxplot(
    #   df, x = "Soil_Type", y = dv, 
    #   color = "Soil_Treatment", palette = "jco") + 
    #   stat_pvalue_manual(pwc_type %>% dplyr::filter(p.adj <= 0.05))
    # 
    # bxp_treatment <- ggboxplot(
    #   df, x = "Soil_Treatment", y = dv, 
    #   color = "Soil_Type", palette = "jco") + 
    #   stat_pvalue_manual(pwc_treatment %>% dplyr::filter(p.adj <= 0.05))
    
    p <- ggbarplot(
      df_sum, x = "Soil_Type", y = "mean", fill = "Soil_Treatment",
      color = "black", position = position_dodge(0.8), size = 1,
      width = 0.8, palette = c("gray99", "gray66", "gray33")) + 
      geom_errorbar(
        aes(
          ymin = ifelse(mean - sd < 0, 0.02, mean - sd),
          ymax = mean + sd), position = position_dodge2(NULL, padding = 0.6), 
        width = 0.8, size = 1) +
      {if(nrow(pwc_type %>% dplyr::filter(p.adj < 0.05, p.adj >= 0.001)) > 0) {
        geom_bracket(
        aes(xmin = xmin, xmax = xmax, label = paste0("list(~italic(p)==", p_round, ")")),
        data = pwc_type %>% dplyr::filter(p.adj < 0.05, p.adj >= 0.001), 
        type = "expression", size = 1, label.size = 3.5, 
        y.position = pwc_type %>% 
          dplyr::filter(p.adj < 0.05, p.adj >= 0.001) %>% 
          dplyr::pull(y.position))}} + 
      {if(nrow(pwc_type %>% dplyr::filter(p.adj < 0.05, p.adj < 0.001)) > 0) {
        geom_bracket(
        aes(xmin = xmin, xmax = xmax, label = "list(~italic(p)<0.001)"),
        data = pwc_type %>% dplyr::filter(p.adj < 0.05, p.adj < 0.001), 
        type = "expression", size = 1, label.size = 3.5, 
        y.position = pwc_type %>% 
          dplyr::filter(p.adj < 0.05, p.adj < 0.001) %>% 
          dplyr::pull(y.position))}} +
      # scale_x_discrete(labels = c(AC = "AC", Ash = "Ash", Con = "Control")) +
      scale_y_continuous(
        expand = c(0, 0), 
        limits = c(0, ifelse(grepl("mass", dv), 4, 1.5)),
        breaks = {if(grepl("mass", dv)) waiver() else seq(0, 1, 0.25)}) +
      ggtitle(ifelse(i == "RF", "Rough Fescue", "Spotted Knapweed")) +
      xlab("Soil Type") +
      ylab(ifelse(dv == "germ_rat", expression(bold(sqrt("Germination Rate"))), 
                  ifelse(dv == "sqrt_mass", expression(bold(sqrt("Biomass (g)"))), "error"))) +
      theme_classic(base_size = 16) +
      theme(
        text = element_text(family = "Times New Roman"),
        legend.position = c(0.18, 0.88),
        legend.background = element_rect(
          fill = "white", size = 1,linetype = "solid", colour = "black"),
        legend.title = element_text(face = "bold", size = 18),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.text = element_text(colour = "black"),
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_line(size = 1, linetype = "solid", colour = "black"),
        axis.title = element_text(face = "bold", size = 16),
        axis.title.y = element_text(margin = margin(t = 0, 10, 0, 0)),
        axis.line = element_line(size = 1), 
        plot.title = element_text(face = "bold", hjust = 0.5)) +
      scale_fill_manual(
        values = c("gray99", "gray66", "gray33"),
        breaks = c("Con", "Ash", "AC"),
        labels = c("Control", "Ash", "Activated Carbon"),
        name = "Soil Treatment:")
    
    return(list(data = df, two_way = aov, tukey = tukey, one_way_type = aov_type,
                one_way_treatment = aov_treatment, pwc_type = pwc_type, 
                pwc_treatment = pwc_treatment, plot = p))
  }, simplify = FALSE, USE.NAMES = TRUE)
}

bmass_analysis <- anova_analysis(do.call(rbind, lapply(bmass_assum, "[[", "data")), "sqrt_mass")

# Prepare data for saving
analysis_input <- lapply(bmass_analysis, "[[", "data") %>% do.call(rbind, .) %>% 
  remove_rownames()

two_way <- lapply(bmass_analysis, function(x) {
  data.frame(x$two_way) %>% 
    dplyr::select(-p..05) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Effect)

tukey <- lapply(bmass_analysis, function(x) {
  data.frame(x$tukey) %>% 
    dplyr::select(-c(p.adj.signif, null.value)) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(term, group1, group2)

anova_type <- lapply(bmass_analysis, function(x) {
  data.frame(x$one_way_type) %>% 
    dplyr::select(-p..05) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Treatment, Effect)

anova_treatment <- lapply(bmass_analysis, function(x) {
  data.frame(x$one_way_treatment) %>% 
    dplyr::select(-p..05) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Type, Effect)

pwc_type <- lapply(bmass_analysis, function(x) {
  x$pwc_type %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y, by = c("Soil_Type", "group1", "group2"),
                                  suffixes = c("", ".y")), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Type, term, .y., group1, group2) %>% 
  dplyr::relocate(starts_with("p.adj"), .after = starts_with("p_")) %>% 
  dplyr::select(-ends_with(".y"))

pwc_treatment <- lapply(bmass_analysis, function(x) {
  x$pwc_treatment %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y, by = c("Soil_Treatment", "group1", "group2"),
                                  suffixes = c("", ".y")), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Treatment, term, .y., group1, group2) %>% 
  dplyr::relocate(starts_with("p.adj"), .after = starts_with("p_")) %>% 
  dplyr::select(-ends_with(".y"))

# In both cases, there is a strong 'Soil_Treatment' effect, and a slight interaction
# effect ('Soil_Type:Soil_Treatment').
# Most of the differences are from using AC in the soils (which killed everything).

# Del Fabbro and Prati's index of soil legacy:

legacy_index <- biomass %>% 
  dplyr::filter(Soil_Treatment != "Ash") %>% 
  group_by(Species, Soil_Type, Soil_Treatment) %>% 
  summarise(mean = mean(TotalMass), .groups = "drop")

legacy <- sapply(unique(biomass$Species), function(x) {
  legacy <- legacy_index %>% 
    dplyr::filter(Species == x) %>% 
    dplyr::select(Soil_Type, Soil_Treatment, mean)
  
  legacy <- (legacy[legacy$Soil_Type == "Invaded" & 
                      legacy$Soil_Treatment == "Con", ]$mean - 
               legacy[legacy$Soil_Type == "Invaded" & 
                        legacy$Soil_Treatment == "AC", ]$mean) - 
    (legacy[legacy$Soil_Type == "Pristine" & 
              legacy$Soil_Treatment == "Con", ]$mean - 
       legacy[legacy$Soil_Type == "Invaded" & 
                legacy$Soil_Treatment == "AC", ]$mean)
}, USE.NAMES = TRUE)

# Write outputs
write.csv(analysis_input, file.path(
  results_dir, "01a_biomass_input_data.csv"), row.names = FALSE)
write.csv(bmass_normal_analysis, file.path(
  results_dir, "01b_biomass_normal_assumptions.csv"), row.names = FALSE)
write.csv(bmass_var_eq_analysis, file.path(
  results_dir, "01c_biomass_var_eq_assumptions.csv"), row.names = FALSE)
write.csv(two_way, file.path(
  results_dir, "01d_biomass_two_way_anova.csv"), row.names = FALSE)
write.csv(tukey, file.path(
  results_dir, "01e_biomass_tukey_results.csv"), row.names = FALSE)
write.csv(anova_type, file.path(
  results_dir, "01f_biomass_one_way_soil_type.csv"), row.names = FALSE)
write.csv(anova_treatment, file.path(
  results_dir, "01g_biomass_one_way_soil_treatment.csv"), row.names = FALSE)
write.csv(pwc_type, file.path(
  results_dir, "01h_biomass_pairwise_comparisons_soil_type.csv"), row.names = FALSE)
write.csv(pwc_treatment, file.path(
  results_dir, "01i_biomass_pairwise_comparisons_soil_treatment.csv"), row.names = FALSE)


```

After the stats, let's make some figures from that data:

```{r biomass figures}

# First, let's make bar charts for the biomass, and afterwards a boxplot, then 
# finally a side-by-side interaction plot. For the bar chart, we need means and
# standard errors of the LogBiomass data, which means creating a new summarized data table
biomass_summary <- analysis_input %>% 
  dplyr::group_by(Soil_Type, Soil_Treatment, Species) %>% 
  dplyr::summarise(mean = mean(sqrt_mass), se = sd(sqrt_mass) / sqrt(n()), .groups = "drop")

#This basically tells us that rough fescue performed best in invaded soils. Odd, but it's a result.
#Now, since I mentioned interactions, I should provide interaction plots for visualization purposes
#I will make the plots and then join them with a common legend afterwards
interactions <- sapply(unique(biomass_summary$Species), function(x) {
  p <- ggplot(subset(biomass_summary, Species == x), 
              aes(Soil_Type, mean, group = Soil_Treatment, linetype = Soil_Treatment)) + 
    geom_line(size = 1) +
    geom_point(size = 2) +
    theme_classic(base_size = 14) +
    theme(
      text = element_text(family = "Times New Roman"),
      legend.position = "none",
      legend.background = element_rect(
        fill = "white",
        size = 1,
        linetype = "solid",
        colour = "black"
      ),
      legend.title = element_text(face = "bold"),
      axis.text = element_text(colour = "black"),
      axis.ticks.x = element_blank(),
      axis.ticks.y = element_line(size = 1, linetype = "solid", colour = "black"),
      axis.title = element_text(face = "bold", size = 16),
      axis.title.y = element_text(margin = margin(t = 0, 10, 0, 0)),
      axis.line = element_line(size = 1),
      plot.title = element_text(face = "bold", hjust = 0.5)
    ) +
    scale_y_continuous(expand = c(0, 0), limits = c(0, 2.5)) +
    scale_x_discrete(labels = c("Invaded", "Pristine")) +
    scale_linetype_manual(
      values = c("solid", "dashed", "dotted"),
      name = "Soil Treatment:",
      breaks = c("Con", "AC", "Ash"),
      labels = c("Control", "Activated Carbon", "Ash")
    ) +
    xlab("Soil Type") +
    ylab(expression(bold(sqrt("Biomass (g)")))) +
    ggtitle(ifelse(x == "RF", "Rough Fescue", "Spotted Knapweed"))
}, simplify = FALSE, USE.NAMES = TRUE)

#Now join those 2 graphs together
final <- ggarrange(plotlist = interactions, common.legend = TRUE, legend = "bottom") +
  ggsave(filename = file.path(plot_dir, "interactions.png"), width = 6, height = 3, dpi = 300)
#Export image at 600px wide x 300px tall

```

Now, plot biomass separated by species. There is also the option of adding an image of the plant to the graphs as well

```{r Biomass separated}

# rf_img <- image_read(
#   "https://live.staticflickr.com/2426/3990317298_1cfbb2e31f_o_d.jpg") %>% 
#   image_rotate(90) %>% 
#   as.raster()
# rf_img_ar <- ncol(rf_img) / nrow(rf_img)
# 
# sk_img <- image_read(
#   "https://upload.wikimedia.org/wikipedia/commons/2/2d/Centaurea_maculosa_Bozeman.jpg") %>% 
#   as.raster()
# sk_img_ar <- ncol(rf_img) / nrow(rf_img)

biomass_plot <- ggarrange(plotlist = list(bmass_analysis$RF$plot,
                                          bmass_analysis$SK$plot), 
                          common.legend = TRUE, legend = "bottom") +
  ggsave(file.path(plot_dir, "bmass.png"), width = 6, height = 3, dpi = 300)

# If images are to be included, run the following lines:
# final <- biomass_plot + 
#   annotation_raster(rf_img, xmin = 0.14, 
#                     xmax = 0.24,
#                     ymin = 0.65,
#                     ymax = 0.88) + 
#   annotation_raster(sk_img, xmin = 0.64, 
#                     xmax = 0.74,
#                     ymin = 0.65,
#                     ymax = 0.88) + 
#   ggsave(file.path(plot_dir, "bmass_with_images.png"), width = 6, height = 3, dpi = 300)

```

I need to do some stuff with survival of the planted greenhouse plants as well:

```{r Survival data load and stats}

# Read in the survival data from the same spreadsheet
surv <- read.csv(file.path(data_dir, "greenhouse_raw.csv"), header = TRUE) %>% 
  dplyr::select(-c(AboveMass, BelowMass, Rep)) %>% 
  dplyr::mutate(Soil_Type = factor(Soil_Type, levels = c("Invaded", "Pristine")),
                Soil_Treatment = factor(Soil_Treatment, levels = c("Con", "AC", "Ash")), 
                germ_rat = Germinants / 5, 
                deaths = ifelse(Germinants == 0, 0, 
                                ((Germinants - Survival) / Germinants)), 
                trial_surv = Survival / 5, 
                germ_surv = ifelse(Germinants == 0, 0, 
                                   Survival / Germinants))

# Create summarized results
surv_summary <- group_by(surv, Soil_Type, Soil_Treatment, Species) %>% 
  summarise(across(c(germ_rat, deaths, trial_surv, germ_surv), 
                   list(mean = mean, se = ~sd(.x) / sqrt(n()))), .groups = "drop")

# Generate list of names that will be used to filter and attribute things later
plots <- names(surv_summary)[endsWith(names(surv_summary), "mean")] %>% 
  gsub("_mean", "", .)

surv_assum <- sapply(plots, function(x) 
  anova_assum(surv, x, rm_outliers = FALSE), simplify = FALSE, USE.NAMES = TRUE)

# Extract results
germ_rat_norm <- lapply(surv_assum$germ_rat, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
death_norm <- lapply(surv_assum$deaths, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
trial_surv_norm <- lapply(surv_assum$trial_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
germ_surv_norm <- lapply(surv_assum$germ_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

# Nothing is normal, sqrt transform and retry

surv_trans <- dplyr::mutate(surv, across(all_of(plots), sqrt))

surv_assum <- sapply(plots, function(x) 
  anova_assum(surv_trans, x, rm_outliers = FALSE), simplify = FALSE, USE.NAMES = TRUE)

# Extract results
germ_rat_norm <- lapply(surv_assum$germ_rat, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
death_norm <- lapply(surv_assum$deaths, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
trial_surv_norm <- lapply(surv_assum$trial_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
germ_surv_norm <- lapply(surv_assum$germ_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

# Fixed germination rate, try other transformations for other variables

surv_trans <- dplyr::mutate(surv, across(all_of(plots[plots != "germ_rat"]), 
                                         ~ log10(max(.x + 1) - .x)), 
                            germ_rat = sqrt(germ_rat))

surv_assum <- sapply(plots, function(x) 
  anova_assum(surv_trans, x, rm_outliers = FALSE), simplify = FALSE, USE.NAMES = TRUE)

# Extract results
death_norm <- lapply(surv_assum$deaths, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
trial_surv_norm <- lapply(surv_assum$trial_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
germ_surv_norm <- lapply(surv_assum$germ_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

# Didn't help anything, try inversing
surv_trans <- dplyr::mutate(surv, across(all_of(plots[plots != "germ_rat"]), 
                                         ~ 1 / (max(.x + 1) - .x)), 
                            germ_rat = sqrt(germ_rat))

surv_assum <- sapply(plots, function(x) 
  anova_assum(surv_trans, x, rm_outliers = FALSE), simplify = FALSE, USE.NAMES = TRUE)

# Extract results
death_norm <- lapply(surv_assum$deaths, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
trial_surv_norm <- lapply(surv_assum$trial_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
germ_surv_norm <- lapply(surv_assum$germ_surv, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

# No simple data transformations worked - use Friedman tests for non-parametric 
# two way ANOVA's -- NOT REALLY CORRECT, USE ALIGNED RANK TRANSFORMATION

germ_rat_analysis <- anova_analysis(do.call(rbind, lapply(surv_assum$germ_rat, "[[", "data")), "germ_rat")

friedman_custom <- function(x, dv) {
  friedman <- sapply(unique(x$Species), function(i) {
    df <- x %>% 
      dplyr::select(-germ_rat) %>% 
      dplyr::filter(Species == i) %>% 
      dplyr::rename(var = sym(dv))
    df2 <- group_by(df, Soil_Type, Soil_Treatment) %>% 
      mutate(Soil_Treatment = paste0(Soil_Treatment, "_", row_number())) %>% 
      ungroup()
    df3 <- group_by(df, Soil_Type, Soil_Treatment) %>% 
      mutate(Soil_Type = paste0(Soil_Type, "_", row_number())) %>% 
      ungroup()
    
    fried_type <- df2 %>% rstatix::friedman_test(var ~ Soil_Treatment | Soil_Type)
    fried_treat <- df3 %>% rstatix::friedman_test(var ~ Soil_Type | Soil_Treatment)
    
    type_fx <- df2 %>% rstatix::friedman_effsize(var ~ Soil_Treatment | Soil_Type)
    treat_fx <- df3 %>% rstatix::friedman_effsize(var ~ Soil_Type | Soil_Treatment)
    
    pwc_type <- df2 %>%
      rstatix::wilcox_test(var ~ Soil_Type, paired = FALSE, p.adjust.method = "bonferroni")
    
    pwc_treatment <- df3 %>%
      rstatix::wilcox_test(var ~ Soil_Treatment, paired = FALSE, p.adjust.method = "bonferroni")
    
    return(list(fried_type = fried_type, fried_treat = fried_treat, type_fx = type_fx,
                treat_fx = treat_fx, pwc_type = pwc_type, pwc_treatment = pwc_treatment))
  }, simplify = FALSE, USE.NAMES = TRUE)
}

death_analysis <- friedman_custom(surv, "deaths")
trial_surv_analysis <- friedman_custom(surv, "trial_surv")
germ_surv_analysis <- friedman_custom(surv, "germ_surv")

# Compare to base ANOVA - did it matter that data wasn't normal etc.?

death_anova <- anova_analysis(surv, "deaths")
trial_surv_anova <- anova_analysis(surv, "trial_surv")
germ_surv_anova <- anova_analysis(surv, "germ_surv")

# Try ARTools package - more complex but more accurate as well
surv_trans <- group_by(surv, Soil_Type, Soil_Treatment, Species) %>% 
  dplyr::mutate(rep = as.factor(paste0("r", row_number()))) %>% 
  ungroup()

art_custom <- function(x, dv) {
  
  sapply(unique(x$Species), function(i) {
    
    # Create data frame to generate stats and figures from for a single species
    df <- x %>% 
      dplyr::select(-any_of("germ_rat")) %>% 
      dplyr::filter(Species == i) %>% 
      dplyr::rename(var = sym(dv)) %>% 
      droplevels() %>% 
      dplyr::mutate(Soil_Type = factor(Soil_Type, levels = c("Invaded", "Pristine")),
                    Soil_Treatment = factor(Soil_Treatment, levels = c("Con", "Ash", "AC"))) %>% 
      as.data.frame() # This is important for ART to run
    
    # Perform aligned rank transformation (ART)
    m <- art(var ~ Soil_Type * Soil_Treatment, data = df)
    anova <- anova(m, response = "art") %>% broom::tidy()
    
    # Note: As of Feb. 2021, there is some new syntax invovled in the ARTool
    # package that might be better to use. Will just comment it out for now
    # since thesis is done, but revisit if necessary. This is the same as running
    # emmeans::contrasts(), so it only produces t.values, not F values:
    main_type <- art.con(m, "Soil_Treatment", method = "eff", adjust = "bonferroni")
    main_trt <- art.con(m, "Soil_Type", method = "eff", adjust = "bonferroni")
    
    # Generate linear model objects of each factor from the main ART model
    m.type <- artlm(m, "Soil_Type")
    m.trt <- artlm(m, "Soil_Treatment")
    
    # Perform analysis of main effects. Function joint_tests() will test the
    # contrasts among EMMs instead of the coefficients. These are most closely
    # related to type III tests and return F values - this was hours of research
    main_type <- joint_tests(m.type, by = "Soil_Treatment") %>% as.data.frame()
    main_trt <- joint_tests(m.trt, by = "Soil_Type") %>% as.data.frame()
    
    title <- case_when(i == "RF" ~ "Rough Fescue",
                       i == "SK" ~ "Spotted Knapweed")
    title <- if(is.na(title)) NULL else title
    
    # Create pairwise comparisons for each factor using emmeans package
    pwc_type_emmeans <- m.type$model %>% 
      group_by(Soil_Treatment) %>% 
      emmeans_test(.y ~ Soil_Type, p.adjust.method = "bonferroni") %>% 
      add_xy_position(x = "Soil_Treatment", fun = "mean_sd", step.increase = 0) %>% 
      dplyr::mutate(y.position = y.position + 8)
    
    pwc_trt_emmeans <- m.trt$model %>% 
      group_by(Soil_Type) %>% 
      emmeans_test(.y ~ Soil_Treatment, p.adjust.method = "bonferroni") %>% 
      add_xy_position(x = "Soil_Type", fun = "mean_sd", step.increase = 0) %>% 
      dplyr::mutate(y.position = y.position + 8)
    
    # Method for adding proper significance labels
    pwc_trt_emmeans <- rbind(
      dplyr::filter(pwc_trt_emmeans, x == 1) %>% 
        dplyr::mutate(y.position = seq(
        from = min(.$y.position), by = 14, length.out = 3)),
      dplyr::filter(pwc_trt_emmeans, x == 2) %>% 
        dplyr::mutate(y.position = seq(
        from = min(.$y.position), by = 14, length.out = 3)))
    
    # Join actual PWC's to the emmeans ones to get proper significance labels
    pwc_trt_join <- dplyr::select(pwc_trt_emmeans, -groups) %>% 
      dplyr::mutate(p_round = round(p.adj, digits = 3))
    
    # Adjust y.position so that everything is tight
    pwc_trt_join <- rbind(
      dplyr::filter(pwc_trt_join, x == 1) %>% 
        {if(dplyr::slice(., which.min(y.position)) %>% 
            dplyr::pull(p.adj) >= 0.05 && any(.$p.adj < 0.05)) {
          tmp <- .
          min_y <- min(tmp$y.position)
          n_inc <- nrow(dplyr::filter(tmp, p.adj < 0.05))
          if(n_inc == 2) {
            sq <- c(min_y, seq(min_y, by = 14, length.out = n_inc))
          } else {
            sq <- rep(min_y, 3)
          }
          dplyr::mutate(., y.position = sq)
        } else . } %>% 
        {if(.$p.adj[2] >= 0.05) {
          dplyr::mutate(., y.position = c(y.position[1], y.position[1], y.position[2]))
        } else . }, 
      dplyr::filter(pwc_trt_join, x == 2) %>% 
        {if(dplyr::slice(., which.min(y.position)) %>% 
            dplyr::pull(p.adj) >= 0.05 && any(.$p.adj < 0.05)) {
          tmp <- .
          min_y <- min(tmp$y.position)
          n_inc <- nrow(dplyr::filter(tmp, p.adj < 0.05))
          if(n_inc == 2) {
            sq <- c(min_y, seq(min_y, by = 14, length.out = n_inc))
          } else {
            sq <- rep(min_y, 3)
          }
          dplyr::mutate(., y.position = sq)
        } else . } %>% 
        {if(.$p.adj[2] >= 0.05) {
          dplyr::mutate(., y.position = c(y.position[1], y.position[1], y.position[2]))
        } else . }
    )
    
    pwc_type_join <- dplyr::select(pwc_type_emmeans, -groups) %>% 
      dplyr::mutate(p_round = round(p.adj, digits = 3)) %>% 
      group_by(x) %>% 
      {if(dplyr::slice(., which.min(y.position)) %>% 
          dplyr::pull(p.adj) > 0.05) {
        dplyr::mutate(., y.position = min(y.position))
          } else .} %>% 
      ungroup()
    
    # Place transformed data back into input df
    df <- dplyr::select(df, Soil_Type, Soil_Treatment, Species, var) %>% 
      dplyr::mutate(soil_type_aligned_ranks = m$aligned.ranks$Soil_Type,
                    soil_trt_aligned_ranks = m$aligned.ranks$Soil_Treatment,
                    all_aligned_ranks = m$aligned.ranks$`Soil_Type:Soil_Treatment`)
    
    # Create dataframes for bar plots
    pwc_trt_data_sum <- m.trt$model %>% 
      group_by(Soil_Type, Soil_Treatment) %>% 
      summarise(across(.y, list(mean = ~mean(.x, na.rm = TRUE),
                                sd = ~sd(.x, na.rm = TRUE)), 
                       .names = "{.fn}"), .groups = "drop")
    
    pwc_type_data_sum <- m.type$model %>% 
      group_by(Soil_Type, Soil_Treatment) %>% 
      summarise(across(.y, list(mean = ~mean(.x, na.rm = TRUE),
                                sd = ~sd(.x, na.rm = TRUE)), 
                       .names = "{.fn}"), .groups = "drop")
    
    # Create bar plots
    p <- ggbarplot(
      pwc_trt_data_sum, x = "Soil_Type", y = "mean", fill = "Soil_Treatment",
      color = "black", position = position_dodge(0.8), size = 1,
      width = 0.8, palette = c("gray99", "gray66", "gray33")) + 
      geom_errorbar(
        aes(
          ymin = ifelse(mean - sd < 0, 2, mean - sd),
          ymax = mean + sd), position = position_dodge2(NULL, padding = 0.6), 
        width = 0.8, size = 1) +
      # Add brackets above bars. Where p > 0.001, use a "=" sign.
      {if(nrow(pwc_trt_join %>% dplyr::filter(p.adj < 0.05, p.adj >= 0.001)) > 0) {
        geom_bracket(
        aes(xmin = xmin, xmax = xmax, label = paste0("list(~italic(p)==", p_round, ")")),
        data = pwc_trt_join %>% dplyr::filter(p.adj < 0.05, p.adj >= 0.001), 
        type = "expression", size = 1, label.size = 3.5)}} + 
      # Add brackets above bars. Where p > 0.001, use a "<" sign.
      {if(nrow(pwc_trt_join %>% dplyr::filter(p.adj < 0.05, p.adj < 0.001)) > 0) {
        geom_bracket(
        aes(xmin = xmin, xmax = xmax, label = "list(~italic(p)<0.001)"),
        data = pwc_trt_join %>% dplyr::filter(p.adj < 0.05, p.adj < 0.001), 
        type = "expression", size = 1, label.size = 3.5)}} + 
      # scale_x_discrete(labels = c(AC = "AC", Ash = "Ash", Con = "Control")) +
      ggtitle(title) +
      xlab("Soil Type") +
      ylab(ifelse(dv == "germ_rat", expression(bold(sqrt("Germination Rate"))), 
                  case_when(
        dv == "deaths" ~ "ART (Death Rate)",
        dv == "trial_surv" ~ "ART (Trial\nSurvival Rate)", 
        dv == "germ_surv" ~ "ART (Germination\nSurvival Rate)",
        dv == "CN" ~ "ART (C:N Ratio)",
        TRUE ~ "error"))) +
      theme_classic(base_size = 16) +
      theme(
        text = element_text(family = "Times New Roman"),
        legend.position = c(0.18, 0.88),
        legend.background = element_rect(
          fill = "white", size = 1,linetype = "solid", colour = "black"),
        legend.title = element_text(face = "bold", size = 18),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        axis.text = element_text(colour = "black"),
        axis.ticks.x = element_blank(),
        axis.ticks.y = element_line(size = 1, linetype = "solid", colour = "black"),
        axis.title = element_text(face = "bold", size = 16),
        axis.title.y = element_text(margin = margin(t = 0, 10, 0, 0)),
        axis.line = element_line(size = 1), 
        plot.title = element_text(face = "bold", hjust = 0.5)) +
      scale_y_continuous(
        expand = c(0, 0), 
        limits = c(0, ifelse(dv == "CN", 40, 100))) +
      scale_fill_manual(
        values = c("gray99", "gray66", "gray33"),
        breaks = c("Con", "Ash", "AC"),
        labels = c("Control", "Ash", "Activated Carbon"),
        name = "Soil Treatment:")
    
    return(list(
      data = df, model = m, two_way = anova, main_type = main_type,
      main_treatment = main_trt, pwc_type = pwc_type_join, 
      pwc_treatment = pwc_trt_join, plot = p))
  }, simplify = FALSE, USE.NAMES = TRUE) 
}

death_art <- art_custom(surv_trans, "deaths")
death_out <- sapply(names(death_art$RF)[!names(death_art$RF) %in% c(
  "model", "plot")], function(x) {
  do.call(rbind, lapply(death_art, "[[", x))}, simplify = FALSE, USE.NAMES = TRUE) %>% 
  lapply(function(x) 
    x %>% 
      rownames_to_column("species") %>% 
      dplyr::select(-any_of("Species")) %>% 
      dplyr::rename(Species = species) %>% 
      dplyr::mutate(Species = substr(Species, 1, 2)))

trial_surv_art <- art_custom(surv_trans, "trial_surv")
trial_surv_out <- sapply(names(trial_surv_art$RF)[!names(death_art$RF) %in% c(
  "model", "plot")], function(x) {
  do.call(rbind, lapply(trial_surv_art, "[[", x))}, simplify = FALSE, USE.NAMES = TRUE) %>% 
  lapply(function(x) 
    x %>% 
      rownames_to_column("species") %>% 
      dplyr::select(-any_of("Species")) %>% 
      dplyr::rename(Species = species) %>% 
      dplyr::mutate(Species = substr(Species, 1, 2)))

germ_surv_art <- art_custom(surv_trans, "germ_surv")
germ_surv_out <- sapply(names(germ_surv_art$RF)[!names(death_art$RF) %in% c(
  "model", "plot")], function(x) {
  do.call(rbind, lapply(germ_surv_art, "[[", x))}, simplify = FALSE, USE.NAMES = TRUE) %>% 
  lapply(function(x) 
    x %>% 
      rownames_to_column("species") %>% 
      dplyr::select(-any_of("Species")) %>% 
      dplyr::rename(Species = species) %>% 
      dplyr::mutate(Species = substr(Species, 1, 2)))

# Write outputs
germ_rat_out <- sapply(names(germ_rat_analysis$RF)[
  !names(germ_rat_analysis$RF) %in% c("plot_type", "plot_trt")], function(x) {
  do.call(rbind, lapply(germ_rat_analysis, "[[", x))}, simplify = FALSE, USE.NAMES = TRUE) %>% 
  lapply(function(x) 
    data.frame(x) %>% 
      rownames_to_column("species") %>% 
      dplyr::select(-any_of("Species")) %>% 
      dplyr::rename(Species = species) %>% 
      dplyr::mutate(Species = substr(Species, 1, 2)))

# Prepare data for saving
germ_rat_analysis_input <- lapply(germ_rat_analysis, "[[", "data") %>%
  do.call(rbind, .) %>% 
  remove_rownames()

germ_rat_two_way <- lapply(germ_rat_analysis, function(x) {
  data.frame(x$two_way) %>% 
    dplyr::select(-p..05) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Effect)

germ_rat_tukey <- lapply(germ_rat_analysis, function(x) {
  data.frame(x$tukey) %>% 
    dplyr::select(-c(p.adj.signif, null.value)) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(term, group1, group2)

germ_rat_anova_type <- lapply(germ_rat_analysis, function(x) {
  data.frame(x$one_way_type) %>% 
    dplyr::select(-p..05) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Treatment, Effect)

germ_rat_anova_treatment <- lapply(germ_rat_analysis, function(x) {
  data.frame(x$one_way_treatment) %>% 
    dplyr::select(-p..05) %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Type, Effect)

germ_rat_pwc_type <- lapply(germ_rat_analysis, function(x) {
  x$pwc_type %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric)) %>% 
    dplyr::rename(!!sym(paste0("p.adj.signif_", x$data$Species[1])) := p.adj.signif)}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Type, term, .y., group1, group2) %>% 
  dplyr::relocate(starts_with("p.adj"), .after = starts_with("p_"))

germ_rat_pwc_treatment <- lapply(germ_rat_analysis, function(x) {
  x$pwc_treatment %>% 
    rename_with(~str_c(., "_", paste0(x$data$Species[1])), where(is.numeric)) %>% 
    dplyr::rename(!!sym(paste0("p.adj.signif_", x$data$Species[1])) := p.adj.signif)}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Treatment, term, .y., group1, group2) %>% 
  dplyr::relocate(starts_with("p.adj"), .after = starts_with("p_"))

germ_rat_norm_assum <- lapply(surv_assum$germ_rat, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")
germ_rat_var_eq_assum <- lapply(surv_assum$germ_rat, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("Species")

write.csv(germ_rat_analysis_input, file.path(
  results_dir, "02a_germ_rat_input_data.csv"), row.names = FALSE)
write.csv(germ_rat_norm_assum, file.path(
  results_dir, "02b_germ_rat_normal_assumption.csv"), row.names = FALSE)
write.csv(germ_rat_var_eq_assum, file.path(
  results_dir, "02c_germ_rat_var_eq_assumption.csv"), row.names = FALSE)
write.csv(germ_rat_two_way, file.path(
  results_dir, "02d_germ_rat_two_way.csv"), row.names = FALSE)
write.csv(germ_rat_tukey, file.path(
  results_dir, "02e_germ_rat_tukey_results.csv"), row.names = FALSE)
write.csv(germ_rat_anova_type, file.path(
  results_dir, "02f_germ_rat_one_way_soil_type.csv"), row.names = FALSE)
write.csv(germ_rat_anova_treatment, file.path(
  results_dir, "02g_germ_rat_one_way_soil_treatment.csv"), row.names = FALSE)
write.csv(germ_rat_pwc_type, file.path(
  results_dir, "02h_germ_rat_pairwise_comparisons_soil_type.csv"), row.names = FALSE)
write.csv(germ_rat_pwc_treatment, file.path(
  results_dir, "02i_germ_rat_pairwise_comparisons_soil_treatment.csv"), row.names = FALSE)

lapply(c("death", "germ_surv", "trial_surv"), function(x) {
  y <- get(paste0(x, "_out"))
  norm_assum <- surv_assum[[ifelse(x == "death", paste0(x, "s"), x)]] %>% 
    lapply("[[", "shapiro_whole") %>% do.call(rbind, .) %>% rownames_to_column("Species")
  var_eq_assum <- surv_assum[[ifelse(x == "death", paste0(x, "s"), x)]] %>% 
    lapply("[[", "levene") %>% do.call(rbind, .) %>% rownames_to_column("Species")
  id <- case_when(x == "death" ~ "03",
                  x == "germ_surv" ~ "04",
                  x == "trial_surv" ~ "05")
  write.csv(norm_assum, file.path(
    results_dir, paste0(id, "a_", x, "_normal_assumption.csv")), row.names = FALSE)
  write.csv(var_eq_assum, file.path(
    results_dir, paste0(id, "b_", x, "_var_eq_assumption.csv")), row.names = FALSE)
  write.csv(y$two_way, file.path(
    results_dir, paste0(id, "c_", x, "_two_way_anova.csv")), row.names = FALSE)
  write.csv(y$main_type, file.path(
    results_dir, paste0(id, "d_", x, "_one_way_soil_type.csv")), row.names = FALSE)
  write.csv(y$main_treatment, file.path(
    results_dir, paste0(id, "e_", x, "_one_way_soil_treatment.csv")), row.names = FALSE)
  write.csv(y$pwc_type, file.path(
    results_dir, paste0(id, "f_", x, "_pairwise_soil_type.csv")), row.names = FALSE)
  write.csv(y$pwc_treatment, file.path(
    results_dir, paste0(id, "g_", x, "_pairwise_soil_treatment.csv")), row.names = FALSE)
})

```

Now plot the survival data

```{r survival plots}

surv_plots <- list(germ_rat = germ_rat_analysis, deaths = death_art, 
                   germ_surv = germ_surv_art, trial_surv = trial_surv_art) %>% 
  lapply(function(x) lapply(x, "[[", "plot"))

# Save individual images
germ_rat_plot <- ggarrange(
  plotlist = unlist(surv_plots["germ_rat"], recursive = FALSE), 
  common.legend = TRUE, legend = "bottom", ncol = 2, nrow = 1) +
  ggsave(filename = file.path(plot_dir, "germ_rate.png"), 
         width = 6, height = 3, dpi = 300)

deaths_plot <- ggarrange(
  plotlist = unlist(surv_plots["deaths"], recursive = FALSE), 
  common.legend = TRUE, legend = "bottom", ncol = 2, nrow = 1) +
  ggsave(filename = file.path(plot_dir, "death_rate.png"), 
         width = 6, height = 3, dpi = 300)

germ_surv_rat_plot <- ggarrange(
  plotlist = unlist(surv_plots["germ_surv"], recursive = FALSE), 
  common.legend = TRUE, legend = "bottom", ncol = 2, nrow = 1) +
  ggsave(filename = file.path(plot_dir, "germ_surv_rate.png"), 
         width = 6, height = 3, dpi = 300)

trial_surv_rat_plot <- ggarrange(
  plotlist = unlist(surv_plots["trial_surv"], recursive = FALSE), 
  common.legend = TRUE, legend = "bottom", ncol = 2, nrow = 1) +
  ggsave(filename = file.path(plot_dir, "trial_surv_rate.png"), 
         width = 6, height = 3, dpi = 300)

# Make a single large plot of all metrics, but remove the titles from all but
# the first metric
surv_plots_no_titles <- c(
  surv_plots[1], lapply(surv_plots[-1], function(x) {
  lapply(names(x), function(y) {
    x[[y]]$labels$title <- ""
    return(x[[y]])
  })
}))

surv_plots_merge <- ggarrange(
  plotlist = unlist(surv_plots_no_titles, recursive = FALSE), 
  common.legend = TRUE, legend = "bottom", ncol = 2, nrow = 4) +
  ggsave(filename = file.path(plot_dir, "survival.png"), 
         width = 8.5, height = 11, dpi = 300)
  

```

Shifting gears to the elemental analyzer data now:

```{r Elemental analyzer data load stats and figure}

analyzer <- read.csv(file.path(data_dir, "Analyzer.csv")) %>% 
  dplyr::mutate(Species = "all", CN = Carbon/Nitrogen) # For tricking my function

el_assum <- lapply(list(
  Nitrogen = anova_assum(analyzer, "Nitrogen", rm_outliers = FALSE), 
  Carbon = anova_assum(analyzer, "Carbon", rm_outliers = FALSE), 
  Hydrogen = anova_assum(analyzer, "Hydrogen", rm_outliers = FALSE), 
  CN = anova_assum(analyzer, "CN", rm_outliers = FALSE)), "[[", "all")

# CN isn't normal/no equal variances, try sqrt transform
analyzer <- dplyr::mutate(analyzer, CN_trans = sqrt(CN))
cn_assum <- anova_assum(analyzer, "CN_trans", rm_outliers = FALSE)

# Didn't work, try log10
analyzer <- dplyr::mutate(analyzer, CN_trans = log10(CN))
cn_assum <- anova_assum(analyzer, "CN_trans", rm_outliers = FALSE)

# Didn't work, try squaring
analyzer <- dplyr::mutate(analyzer, CN_trans = CN^2)
cn_assum <- anova_assum(analyzer, "CN_trans", rm_outliers = FALSE)

# Didn't work, try inverse
analyzer <- dplyr::mutate(analyzer, CN_trans = 1/CN)
cn_assum <- anova_assum(analyzer, "CN_trans", rm_outliers = FALSE)

# Didn't work, will have to do non-parametric 2 way ANOVA
analyzer <- dplyr::select(analyzer, -CN_trans)

# All assumptions are validated, good to use ANOVA
el_anova <- c(lapply(el_assum[c(1:3)], function(x) 
  anova_analysis(x$data, names(x$data)[4])[[1]]),
  CN = unname(art_custom(analyzer, "CN")))

# Rename columns to match other names for export
analyzer$CN <- el_anova$CN$data$all_aligned_ranks
el_anova$CN$data <- dplyr::rename(el_anova$CN$data, CN = var)
el_anova$CN$two_way <- el_anova$CN$two_way %>% 
  dplyr::rename(Effect = term, DFn = df, DFd = Df.res,
                `F` = statistic, p = p.value) %>% 
  dplyr::select(-c(sumsq, Sum.Sq.res))

el_anova$CN$main_type <- el_anova$CN$main_type %>% 
  dplyr::rename(Effect = `model term`, DFn = df1, DFd = df2,
                `F` = F.ratio, p = p.value)

el_anova$CN$main_treatment <- el_anova$CN$main_treatment %>% 
  dplyr::rename(Effect = `model term`, DFn = df1, DFd = df2,
                `F` = F.ratio, p = p.value)

names(el_anova$CN) <- gsub("main", "one_way", names(el_anova$CN))
names(el_anova$CN) <- gsub("pwc_type", "pwc_treatment_", names(el_anova$CN))
names(el_anova$CN) <- gsub("pwc_treatment$", "pwc_type", names(el_anova$CN))
names(el_anova$CN) <- gsub("pwc_treatment_", "pwc_treatment", names(el_anova$CN))

# Prepare data for saving
el_norm_assum <- lapply(el_assum, "[[", "shapiro_whole") %>% 
  do.call(rbind, .) %>% rownames_to_column("Element")
el_var_eq_assum <- lapply(el_assum, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("Element")

el_analysis_input <- lapply(el_anova, "[[", "data") %>% 
  lapply(function(x) pivot_longer(x, cols = c("Nitrogen", "Carbon", "Hydrogen", "CN")[
    c("Nitrogen", "Carbon", "Hydrogen", "CN") %in% names(x)], names_to = "Element", 
    values_to = "total") %>% 
      dplyr::select(Soil_Type, Soil_Treatment, Element, total)) %>% 
  do.call(rbind, .) %>% 
  remove_rownames()

el_analysis_two_way <- lapply(el_anova, function(x) {
  data.frame(x$two_way) %>% 
    dplyr::select(-any_of(c("Term", "p..05", "ges"))) %>% 
    rename_with(~str_c(., "_", paste0(names(x$data)[4])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Effect)

el_analysis_tukey <- lapply(el_anova[1:3], function(x) {
  data.frame(x$tukey) %>% 
    dplyr::select(-c(p.adj.signif, null.value)) %>% 
    rename_with(~str_c(., "_", paste0(names(x$data)[4])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(term, group1, group2)

el_analysis_anova_type <- lapply(el_anova, function(x) {
  data.frame(x$one_way_type) %>% 
    dplyr::select(-any_of(c("p..05", "ges"))) %>% 
    rename_with(~str_c(., "_", paste0(names(x$data)[4])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Treatment, Effect)

el_analysis_anova_treatment <- lapply(el_anova, function(x) {
  data.frame(x$one_way_treatment) %>% 
    dplyr::select(-any_of(c("p..05", "ges"))) %>% 
    rename_with(~str_c(., "_", paste0(names(x$data)[4])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Type, Effect)

el_analysis_pwc_type <- lapply(el_anova, function(x) {
  x$pwc_type %>% 
    dplyr::select(-any_of(c("df", ".y.", "p.adj.signif"))) %>% 
    rename_with(~str_c(., "_", paste0(names(x$data)[4])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y, by = c("Soil_Type", "term", "group1", "group2")), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Type, term, group1, group2) %>% 
  dplyr::relocate(starts_with("p.adj"), .after = starts_with("p_"))

el_analysis_pwc_treatment <- lapply(el_anova, function(x) {
  x$pwc_treatment %>% 
    dplyr::select(-any_of(c("df", ".y.", "p.adj.signif"))) %>% 
    rename_with(~str_c(., "_", paste0(names(x$data)[4])), where(is.numeric))}) %>% 
  Reduce(f = function(x, y) merge(x, y, by = c("Soil_Treatment", "term", "group1", "group2")), x = .) %>% 
  dplyr::select(sort(tidyselect::peek_vars())) %>% 
  dplyr::relocate(Soil_Treatment, term, group1, group2) %>% 
  dplyr::relocate(starts_with("p.adj"), .after = starts_with("p_"))

# Create summary data for plotting
# analyzer_summary <- group_by(analyzer, Soil_Type, Soil_Treatment) %>% 
#   dplyr::summarise(across(where(is.numeric), list(mean = mean, se = ~sd(.x) / sqrt(n()))), .groups = "drop")

# Write outputs
write.csv(el_analysis_input, file.path(
  results_dir, "06a_elemental_analysis_input_data.csv"), row.names = FALSE)
write.csv(el_norm_assum, file.path(
  results_dir, "06b_elemental_analysis_normal_assumption.csv"), row.names = FALSE)
write.csv(el_var_eq_assum, file.path(
  results_dir, "06c_elemental_analysis_var_eq_assumption.csv"), row.names = FALSE)
write.csv(el_analysis_two_way, file.path(
  results_dir, "06d_elemental_analysis_two_way.csv"), row.names = FALSE)
write.csv(el_analysis_tukey, file.path(
  results_dir, "06e_elemental_analysis_tukey_results.csv"), row.names = FALSE)
write.csv(el_analysis_anova_type, file.path(
  results_dir, "06f_elemental_analysis_one_way_soil_type.csv"), row.names = FALSE)
write.csv(el_analysis_anova_treatment, file.path(
  results_dir, "06g_elemental_analysis_one_way_soil_treatment.csv"), row.names = FALSE)
write.csv(el_analysis_pwc_type, file.path(
  results_dir, "06h_elemental_analysis_pairwise_comparisons_soil_type.csv"), row.names = FALSE)
write.csv(el_analysis_pwc_treatment, file.path(
  results_dir, "06i_elemental_analysis_pairwise_comparisons_soil_treatment.csv"), row.names = FALSE)

# Plot results
analyzer_plots <- sapply(names(el_anova)[1:3], function(x) {
  df <- dplyr::select(analyzer, c(Soil_Type, Soil_Treatment, all_of(x))) %>% 
    unite(trt, c("Soil_Type", "Soil_Treatment"), remove = FALSE) %>% 
    droplevels() %>% 
    dplyr::mutate(Soil_Treatment = factor(Soil_Treatment, levels = c("Con", "Ash", "AC"))) %>% 
    dplyr::rename(data = 4)
  
  pwc <- el_anova[[x]]$pwc_type
  # if(x == "CN") {
  #   pwc <- dplyr::mutate(pwc, y.position = y.position - 18.5)
  # }
  
  label <- if(x == "CN") "ART (C:N Ratio)" else paste0("Total ", x, " (%)")
  
  p <- ggboxplot(df, x = "Soil_Type", y = "data", fill = "Soil_Treatment", 
                 width = 0.7, color = "black", palette = c("gray99", "gray66", "gray33"),
                 size = 0.5, fatten = 1) +
    # Add brackets above bars. Where p > 0.001, use a "=" sign.
    {if(nrow(pwc %>% dplyr::filter(p.adj < 0.05, p.adj >= 0.001)) > 0) {
      geom_bracket(
        aes(xmin = xmin, xmax = xmax, label = paste0("list(~italic(p)==", p_round, ")")),
        data = pwc %>% dplyr::filter(p.adj < 0.05, p.adj >= 0.001), 
        type = "expression", size = 1, label.size = 3.5)}} + 
    # Add brackets above bars. Where p > 0.001, use a "<" sign.
    {if(nrow(pwc %>% dplyr::filter(p.adj < 0.05, p.adj < 0.001)) > 0) {
      geom_bracket(
        aes(xmin = xmin, xmax = xmax, label = "list(~italic(p)<0.001)"),
        data = pwc %>% dplyr::filter(p.adj < 0.05, p.adj < 0.001), 
        type = "expression", size = 1, label.size = 3.5)}} +
    xlab("Soil Type") +
    ylab(label)  +
    theme_classic(base_size = 16) +
    theme(
      text = element_text(family="Times New Roman"), 
      legend.position = c(0.18, 0.88),
      legend.background = element_rect(
        fill = "white", size = 1, linetype = "solid", colour = "black"),
      legend.title = element_text(face = "bold",size = 16), 
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      axis.text = element_text(colour = "black"),
      axis.ticks.x = element_blank(),
      axis.ticks.y = element_line(size = 1, linetype = "solid", colour = "black"),
      axis.title = element_text(face = "bold",size = 16),
      axis.title.y = element_text(margin = margin(t = 0, 10, 0, 0)),
      axis.line = element_line(size = 1)) +
    scale_fill_manual(
      values = c("gray99", "gray66", "gray33"),
      breaks = c("Con", "Ash", "AC"),
      labels = c("Control", "Ash", "Activated Carbon"), 
      name = "Soil Amendment:") +
    scale_x_discrete(labels = c(INV = "Invaded", PRI = "Pristine")) +
    scale_y_continuous(
      expand = c(0, 0), 
      limits = c(
        ifelse(max(df$data) < 1 | max(df$data) - min(df$data) < 1, 
               round(min(df$data) - 0.05, digits = 2), min(floor(df$data))),
        ifelse(max(df$data) < 1 | max(df$data) - min(df$data) < 1, 
               round(max(df$data) + 0.15, digits = 2), max(ceiling(df$data)))))
  
}, simplify = FALSE, USE.NAMES = TRUE)

# Arrange final plot of carbon and nitrogen only
final <- ggarrange(plotlist = analyzer_plots[2:1], common.legend = TRUE, legend = "bottom") +
  ggsave(filename = file.path(plot_dir, "CN_plot.png"), width = 6, height = 3, dpi = 300)

final2 <- ggarrange(plotlist = list(analyzer_plots[[3]], el_anova$CN$plot), common.legend = TRUE, legend = "bottom") +
  ggsave(filename = file.path(plot_dir, "H_CN_plot.png"), width = 6, height = 3, dpi = 300)

final3 <- ggarrange(plotlist = c(
  analyzer_plots[2:1], list(analyzer_plots[[3]], el_anova$CN$plot)), common.legend = TRUE, legend = "bottom", ncol = 2, nrow = 2) +
  ggsave(filename = file.path(plot_dir, "EL_Analysis_plot.png"), width = 6, height = 6, dpi = 300)


```
