==> 03_Field_Surveys.Rmd <==
---
title: "03_Field_Surveys"
author: "Matt"
date: "9/15/2020"
output: html_document
---

This document will go through the field surveys from each field plot and produce appropriate diversity indices and other interesting metrics.

```{r Load packages}

suppressMessages(suppressWarnings({
  ls <- c("tidyverse", "broom", "car", "ggpubr", "car", "sf", 
          "extrafont", "vegan", "rstatix", "coin")
  new_packages <- ls[!(ls %in% installed.packages()[, "Package"])]
  if(length(new_packages)) install.packages(new_packages)
  sapply(ls, library, character.only = TRUE, quietly = TRUE)[0]
  if(any(!fonts() %in% unlist(windowsFonts())) || length(fonts()) == 0) {
    font_import(prompt = FALSE)
    loadfonts(device = "win")}
  rm(ls, new_packages)}))

```

Load data

```{r Set directories load plant cover data}

# Set directories
data_dir <- file.path("./03_Field_Surveys/data")
plot_dir <- file.path("./03_Field_Surveys/plots")
results_dir <- file.path("./03_Field_Surveys/results")
dir.create(plot_dir, showWarnings = FALSE)
dir.create(results_dir, showWarnings = FALSE)

# Read in plant cover data
cov <- read.csv(file.path(data_dir, "plant_data.csv")) %>% 
  separate(Site_ID, c("Site", "Rep"), remove = FALSE) %>% 
  mutate(Rep = as.numeric(Rep))

# In order to use the diversity function from the vegan package, the data need
# to be pivoted where each row is a separate site and columns are plant species
cov_pivot <- sapply(unique(cov$Site), function(x) {
  dplyr::filter(cov, Site == x) %>% 
    pivot_wider(Site_ID, names_from = Common.Name, values_from = Cover) %>% 
    replace(is.na(.), 0)
}, simplify = FALSE, USE.NAMES = TRUE)

# Perform diversity and richness calculations separately for each site
shannon <- lapply(cov_pivot, function(x) 
  data.frame(
    Site_ID = x$Site_ID, 
    shannon = diversity(x %>% dplyr::select(-Site_ID), index = "shannon")))
simpson <- lapply(cov_pivot, function(x) 
  data.frame(
    Site_ID = x$Site_ID, 
    simpson = diversity(x %>% dplyr::select(-Site_ID), index = "simpson")))
richness <- lapply(cov_pivot, function(x) 
  data.frame(
    Site_ID = x$Site_ID, 
    richness = apply(x %>% dplyr::select(-Site_ID), MARGIN = 1, function(y) sum(y != 0))))

# Bring the calculations back together in a single table
cov_summary <- lapply(list(shannon = shannon, simpson = simpson, richness = richness), function(x)
  do.call(rbind, x) %>% 
    pivot_longer(names(.)[2], names_to = "metric")) %>% 
  do.call(rbind, .) %>% 
  separate(Site_ID, c("Site", "Rep"), remove = FALSE) %>% 
  dplyr::mutate(Rep = as.numeric(Rep), 
                metric = factor(metric, levels = c("shannon", "simpson", "richness")), 
                Site = as.factor(Site))

cov_sum_stats <- cov_summary %>% 
  group_by(Site, metric) %>% 
  get_summary_stats(value)

# Want to perform t-test, so perform assumptions starting with if variables are 
# equal (shapiro test), and then if the data are normal 
ttest_assum <- function(x, dv, rm_outliers = TRUE) {
  sapply(unique(as.character(x$metric)), function(i) {
    df <- dplyr::filter(x, metric == i) %>% 
      dplyr::rename(var = sym(dv)) %>% 
      group_by(Site) %>% 
      mutate(outlier = is_outlier(var), extreme = is_extreme(var)) %>% 
      ungroup() %>% 
      {if(rm_outliers) dplyr::filter(., !extreme) else .}
    shapiro_model <- df %>% 
      group_by(Site) %>% 
      shapiro_test(var)
    norm_plot_model <- ggqqplot(df, x = "var", facet.by = "Site")
    levene <- df %>% levene_test(var ~ Site)
    df <- dplyr::rename(df, !!dv := var)
    return(list(data = df, shapiro = shapiro_model, 
                norm_plot = norm_plot_model, levene = levene))
  }, simplify = FALSE, USE.NAMES = TRUE)
}

cov_assum <- ttest_assum(cov_summary, dv = "value", rm_outliers = FALSE)
cov_norm_assum <- lapply(cov_assum, "[[", "shapiro") %>% do.call(rbind, .) %>% 
  rownames_to_column("metric") %>% 
  dplyr::mutate(metric = gsub("\\..*", "", metric))
cov_var_eq_assum <- lapply(cov_assum, "[[", "levene") %>% do.call(rbind, .) %>% 
  rownames_to_column("metric")

# All tests come back showing that the data are normal and have equal variances
# allowing us to proceed with a t-test
ttest_custom <- function(x, dv, var.equal = TRUE) {
  sapply(unique(as.character(x$metric)), function(i) {
    df <- dplyr::filter(x, metric == i) %>% 
      dplyr::rename(var = sym(dv))
    ttest <- df %>% t_test(var ~ Site, var.equal = var.equal) %>% 
      add_significance()
    ttest_plot <- ttest %>% add_xy_position(x = "Site")
    effect <- df %>% cohens_d(var ~ Site, var.equal = var.equal)
    plot <- ggboxplot(
      df, x = "Site", y = "var", ylab = dv, xlab = "Site", add = "jitter") +
      stat_pvalue_manual(ttest_plot, tip.length = 0) + 
      labs(subtitle = get_test_label(ttest_plot, detailed = TRUE))
    df <- dplyr::rename(df, !!dv := var)
    return(list(data = df, ttest = ttest, effect = effect, plot = plot))
  }, simplify = FALSE, USE.NAMES = TRUE)
}

cov_analysis <- lapply(cov_assum, "[[", "data") %>% do.call(rbind, .) %>% 
  ttest_custom(dv = "value", var.equal = TRUE)

cov_ttest <- lapply(cov_analysis, "[[", "ttest") %>% do.call(rbind, .) %>% 
  rownames_to_column("metric")
cov_effect <- lapply(cov_analysis, "[[", "effect") %>% do.call(rbind, .) %>% 
  rownames_to_column("metric")

# Species richness is not significantly different between sites, but in the field
# the invaded sites look much poorer than the pristine sites. Perhaps if richness
# was split by invasive vs. native plant types, that could tease some interesting
# results. I'll use the invasive plant list generated from the IAPP database

# Copied code from 01_Spatial_Analysis


curl::curl_download(
  "https://www2.gov.bc.ca/assets/gov/environment/plants-animals-and-ecosystems/invasive-species/iapp-resources/iapp_species_list_and_map_codes_feb2020.xlsx", 
  destfile = file.path(data_dir, "iapp_species_list_and_map_codes.xlsx"))

bc_iapp_plant_table <- readxl::read_xlsx(
  file.path(data_dir, "iapp_species_list_and_map_codes.xlsx"), col_names = TRUE) %>% 
  rename_with(make.names) %>% 
  dplyr::arrange(Map.Symbol)

# Merge together to find which records are of invasive plant species
cov_invasive <- rbind(
  merge(bc_iapp_plant_table, cov, by.x = "Latin.Name", by.y = "Scientific.Name") %>% 
    dplyr::select(Latin.Name, Site_ID) %>% 
    dplyr::rename(Scientific.Name = Latin.Name),
  merge(bc_iapp_plant_table, cov, by = "Common.Name") %>% 
    dplyr::select(Scientific.Name, Site_ID)
) %>% dplyr::mutate(Invasive = TRUE) %>% 
  merge(cov, by = c("Scientific.Name", "Site_ID"), all = TRUE) %>% 
  replace_na(list(Invasive = FALSE))

# Extract names of all recorded invasive plants in my study site
invasive_plants <- dplyr::filter(cov_invasive, Invasive) %>% 
  dplyr::pull(Common.Name) %>% unique()

# Create separate pivot tables - one for invasive plants, and one for native 
# plants, both of which have recorded cover values at invaded and pristine sites
cov_invasive_pivot <- sapply(unique(cov_invasive$Site), function(x) {
  dplyr::filter(cov_invasive, Site == x) %>% 
    pivot_wider(Site_ID, names_from = Common.Name, values_from = Cover) %>% 
    dplyr::select(any_of(invasive_plants), Site_ID) %>% 
    replace(is.na(.), 0)
}, simplify = FALSE, USE.NAMES = TRUE)

cov_native_pivot <- sapply(unique(cov_invasive$Site), function(x) {
  dplyr::filter(cov_invasive, Site == x) %>% 
    pivot_wider(Site_ID, names_from = Common.Name, values_from = Cover) %>% 
    dplyr::select(-any_of(invasive_plants)) %>%
    replace(is.na(.), 0)
}, simplify = FALSE, USE.NAMES = TRUE)

# Calculate richness for each of those
invasive_richness <- lapply(cov_invasive_pivot, function(x) 
  data.frame(
    Site_ID = x$Site_ID,
    invasive_richness = apply(x %>% dplyr::select(-Site_ID), MARGIN = 1, function(y) 
      sum(y != 0))))

native_richness <- lapply(cov_native_pivot, function(x)
  data.frame(
    Site_ID = x$Site_ID,
    native_richness = apply(x %>% dplyr::select(-Site_ID), MARGIN = 1, function(y) 
      sum(y != 0))))

# Combine the lists and create tables for invasive plant richness and native plant 
# richness
richness_ls <- lapply(
  list(inv_richness = invasive_richness, nat_richness = native_richness),
  function(x)
    do.call(rbind, x) %>% 
    separate(Site_ID, c("Site", "Rep"), remove = FALSE) %>% 
    pivot_longer(names(.)[4], names_to = "metric") %>% 
    dplyr::mutate(Rep = as.numeric(Rep), 
                  Site = as.factor(Site), 
                  metric = as.factor(metric))) 

# The new separated richness tables are ready for analysis (t-tests) - test assumptions
inv_richness_assum_raw <- lapply(richness_ls, ttest_assum, dv = "value", rm_outliers = FALSE) %>% 
  lapply("[[", 1)
inv_richness_norm_assum_raw <- lapply(inv_richness_assum_raw, "[[", "shapiro") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness") %>% 
  dplyr::mutate(sp_richness = gsub("\\..*", "", sp_richness))
inv_richness_var_eq_assum_raw <- lapply(inv_richness_assum_raw, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness")

# Data not normal, try square root transforming. NOTE: bad variable naming here,
# but it works...
nat_richness_assum <- lapply(richness_ls, mutate, sqrt_val = sqrt(value)) %>% 
  lapply(ttest_assum, dv = "sqrt_val", rm_outliers = FALSE) %>% lapply("[[", 1)
nat_richness_norm_assum <- lapply(nat_richness_assum, "[[", "shapiro") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness") %>% 
  dplyr::mutate(sp_richness = gsub("\\..*", "", sp_richness))
nat_richness_var_eq_assum <- lapply(nat_richness_assum, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness")

# Native richness good, try log transforming for invasive richness
inv_richness_assum <- lapply(richness_ls["inv_richness"], mutate, log_val = log10(value + 1)) %>% 
  lapply(ttest_assum, dv = "log_val", rm_outliers = FALSE) %>% lapply("[[", 1)
inv_richness_norm_assum <- lapply(inv_richness_assum, "[[", "shapiro") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness") %>% 
  dplyr::mutate(sp_richness = gsub("\\..*", "", sp_richness))
inv_richness_var_eq_assum <- lapply(inv_richness_assum, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness")

# No good, try inverse transformation
inv_richness_assum <- lapply(richness_ls["inv_richness"], mutate, inv_val = 1 / (max(value + 1) - value)) %>% 
  lapply(ttest_assum, dv = "inv_val", rm_outliers = FALSE) %>% lapply("[[", 1)
inv_richness_norm_assum <- lapply(inv_richness_assum, "[[", "shapiro") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness") %>% 
  dplyr::mutate(sp_richness = gsub("\\..*", "", sp_richness))
inv_richness_var_eq_assum <- lapply(inv_richness_assum, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("sp_richness")

# No transformation helped with invasive plant richness, use non-parametric version
# Wilcoxon rank sum test (AKA Mann_Whitney)

nat_richness_analysis <- ttest_custom(nat_richness_assum$nat_richness$data, 
                                      dv = "sqrt_val", var.equal = TRUE)[[1]]
inv_richness_analysis <- list(
  data = inv_richness_assum$inv_richness$data,
  ttest = inv_richness_assum$inv_richness$data %>% 
    rstatix::wilcox_test(value ~ Site) %>% 
    add_significance(),
  effect = inv_richness_assum$inv_richness$data %>% 
    wilcox_effsize(value ~ Site))
inv_richness_analysis$plot <- ggboxplot(
  inv_richness_analysis$data, x = "Site", y = "value", ylab = "Richness", 
  xlab = "Site", add = "jitter") + 
  stat_pvalue_manual(inv_richness_analysis$ttest %>% add_xy_position(x = "Site"),
                     tip.length = 0) +
  labs(subtitle = get_test_label(inv_richness_analysis$ttest, detailed = TRUE))

# Separating native species richness vs. invasive species richness at each site
# yields significantly different species richnesses. This data should be added
# to the cov_summary dataframe
cov_summary2 <- bind_rows(cov_summary, do.call(rbind, richness_ls)) %>% droplevels()

# Prepare data for output
shapiro_out <- bind_rows(
  nat_richness_assum$nat_richness$shapiro %>% dplyr::mutate(metric = "native_richness"), 
  inv_richness_assum_raw$inv_richness$shapiro %>% dplyr::mutate(metric = "invasive_richness"),
  cov_norm_assum)

levene_out <- bind_rows(
  nat_richness_assum$nat_richness$levene %>% dplyr::mutate(metric = "native_richness"), 
  inv_richness_assum_raw$inv_richness$levene %>% dplyr::mutate(metric = "invasive_richness"),
  cov_var_eq_assum)

ttest_out <- bind_rows(
  nat_richness_analysis$ttest %>% dplyr::mutate(metric = "native_richness"),
  inv_richness_analysis$ttest %>% dplyr::mutate(metric = "invasive_richness"),
  do.call(rbind, lapply(cov_analysis, "[[", "ttest")) %>% rownames_to_column("metric"))

effects_out <- rbind(
  data.frame(nat_richness_analysis$effect) %>% dplyr::mutate(metric = "native_richness"),
  data.frame(inv_richness_analysis$effect) %>% dplyr::mutate(metric = "invasive_richness"),
  data.frame(do.call(rbind, lapply(cov_analysis, "[[", "effect"))) %>%
    rownames_to_column("metric"))


# Write some outputs
write.csv(cov_summary2, file.path(
  results_dir, "01a_diversity_calculations.csv"), row.names = FALSE)
write.csv(shapiro_out, file.path(
  results_dir, "01b_diversity_normal_assumption.csv"), row.names = FALSE)
write.csv(levene_out, file.path(
  results_dir, "01c_diversity_var_equal_assumption.csv"), row.names = FALSE)
write.csv(ttest_out, file.path(
  results_dir, "01d_diversity_ttests.csv"), row.names = FALSE)
write.csv(effects_out, file.path(
  results_dir, "01e_diversity_effects.csv"), row.names = FALSE)

```

The data is mostly prepared. Now to focus on plotting it in a boxplot

```{r Plant diversity plot}

# This is code to generate a 3 paneled figure with 2 y-axes. It looked ok but
# wasn't my favorite.
# p <- ggplot(cov_summary %>% 
#               mutate(value = ifelse(metric == "richness", value / 7, value)), 
#             aes(x = Site, y = value)) + 
#   geom_boxplot(aes(fill = Site), color = "black") +
#   theme_classic(base_size = 14) +
#   theme(
#     rect = element_rect(fill = "transparent"),
#     text = element_text(family = "Times New Roman"),
#     legend.position = "bottom",
#     legend.background = element_rect(
#       fill = "white",
#       size = 1,
#       linetype = "solid",
#       colour = "black"),
#     legend.title = element_text(face = "bold", size = 16),
#     panel.grid.major = element_blank(),
#     panel.grid.minor = element_blank(),
#     axis.text = element_text(colour = "black"),
#     axis.text.x = element_blank(),
#     axis.ticks.x = element_blank(),
#     axis.ticks.y = element_line(size = 1, linetype = "solid", colour = "black"),
#     axis.title = element_text(face = "bold", size = 16),
#     axis.title.x = element_blank(),
#     axis.title.y = element_text(margin = margin(t = 0, 10, 0, 0)),
#     axis.title.y.right = element_text(angle = 90),
#     axis.line = element_blank(),
#     plot.title = element_text(face = "bold", hjust = 0.5), 
#     panel.border = element_rect(colour = "black", size = 2), 
#     strip.background = element_blank(), 
#     strip.text = element_text(face = "bold", size = 14)) +
#   facet_wrap(
#     ~ metric, ncol = 3, strip.position = "bottom",
#     labeller = labeller(
#       metric = c("Shannon (H)", "Simpson (D)", "Richness (S)") %>% 
#         magrittr::set_names(c("shannon", "simpson", "richness")))) +
#   scale_y_continuous(
#     expand = c(0, 0),
#     limits = c(0, round(max(
#       cov_summary[cov_summary$metric != "richness", ]$value), digits = 1) + 0.3),
#     
#     # Features of the first axis
#     name = "Shannon (H) and Simpson (D)\nDiversity Indices",
#     
#     # Add a second axis and specify its features
#     sec.axis = sec_axis(~ . * 7, name = "Species Richness (S)")) +
#   scale_fill_manual(
#     values = c("gray99", "gray66"),
#     breaks = c("Inv", "Pri"),
#     labels = c("Invaded", "Pristine"), 
#     name = "Site:") + 
#   stat_compare_means(
#     method = "t.test", method.args = list(var.equal = TRUE), label = "p.signif", 
#     hide.ns = TRUE, label.x.npc = "center", size = 10, fontface = "bold", 
#     label.y = round(max(
#       cov_summary[cov_summary$metric != "richness", ]$value), digits = 1)) +
#   ggsave(file.path(plot_dir, "diversity.png"), width = 12, height = 4, dpi = 300)

plots <- lapply(c("diversity", "richness"), function(x) {
  flt <- unlist(ifelse(
    x == "diversity", list(c("shannon", "simpson")), 
    list(c("native_richness", "invasive_richness"))))
  p <- ggplot(cov_summary2 %>% dplyr::filter(metric %in% flt), aes(x = Site, y = value)) +
    geom_boxplot(aes(fill = Site), color = "black") +
    theme_classic(base_size = 14) +
    ylab(ifelse(x == "diversity", "Diversity Index Score", "Species Richness Score")) +
    theme(
      rect = element_rect(fill = "transparent"),
      text = element_text(family = "Times New Roman"),
      legend.position = "bottom",
      legend.background = element_rect(
        fill = "white",
        size = 1,
        linetype = "solid",
        colour = "black"),
      legend.title = element_text(face = "bold", size = 16),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      axis.text = element_text(colour = "black"),
      axis.text.x = element_blank(),
      axis.ticks.x = element_blank(),
      axis.ticks.y = element_line(size = 1, linetype = "solid", colour = "black"),
      axis.title = element_text(face = "bold", size = 16),
      axis.title.x = element_blank(),
      axis.title.y = element_text(margin = margin(t = 0, 20, 0, 0)),
      axis.line = element_blank(),
      panel.border = element_rect(colour = "black", size = 2),
      strip.background = element_rect(colour = "black", size = 2),
      strip.text = element_text(face = "bold", size = 14),
      plot.title = element_text(face = "bold", hjust = 0.5)) +
    scale_y_continuous(
      expand = c(0, 0),
      limits = c(0, ifelse(x == "diversity", 2.5, 15))) +
    scale_fill_manual(
      values = c("gray99", "gray66"),
      breaks = c("Inv", "Pri"),
      labels = c("Invaded", "Pristine"),
      name = "Site:") +
    facet_wrap(
      ~ metric, ncol = 2, strip.position = "top",
      labeller = labeller(
        metric = unlist(ifelse(x == "diversity", list(c("Shannon (H)", "Simpson (D)")), 
                        list(c("Native Richness (S)", "Invasive Richness (S)")))) %>%
          magrittr::set_names(flt))) +
    stat_compare_means(
      method = "t.test", method.args = list(var.equal = TRUE), label = "p.signif",
      hide.ns = TRUE, label.x.npc = "center", size = 10, fontface = "bold", 
      label.y = ifelse(x == "diversity", 2.1, 12.6))
})

# Combine
diversity_combined <- ggarrange(plotlist = plots, ncol = 1, common.legend = TRUE, legend = "bottom") +
  ggsave(file.path(plot_dir, "diversity.png"), width = 6, height = 6, dpi = 300)

```

The above analysis only used plant data, however bare ground and litter cover values were also collected in the field. I'll do some analysis on those now too:

```{r Litter analysis}

litter <- read.csv(file.path(data_dir, "litter_bg_data.csv"), header = TRUE) %>% 
  pivot_wider(Site_ID, names_from = Common.Name, values_from = Cover) %>% 
  replace(is.na(.), 0) %>% 
  separate(Site_ID, c("Site", "Rep"), remove = FALSE) %>% 
  mutate(Rep = as.numeric(Rep), Site = as.factor(Site))

litter_sum_stats <- litter %>% 
  group_by(Site) %>% 
  get_summary_stats(Litter)

bg_sum_stats <- litter %>% 
  group_by(Site) %>% 
  get_summary_stats(`Bare Ground`)

litter_assumptions <- sapply(c("Litter", "Bare Ground"), function(x) {
  df <- dplyr::rename(litter, data = all_of(x))
  shapiro <- df %>% group_by(Site) %>% shapiro_test(data)
  levene <- df %>% levene_test(data ~ Site)
  return(list(shapiro = shapiro, levene = levene))
}, simplify = FALSE, USE.NAMES = TRUE)

litter_normal_analysis <- lapply(litter_assumptions, "[[", "shapiro") %>% 
  do.call(rbind, .) %>% rownames_to_column("metric") %>%
  dplyr::mutate(metric = gsub("\\..*", "", metric))
  
litter_var_eq_analysis <- lapply(litter_assumptions, "[[", "levene") %>% 
  do.call(rbind, .) %>% rownames_to_column("metric") 

# Litter data passes checks, bare ground data does not. Transform bare ground
# data and try again
bg <- dplyr::mutate(litter, bg_trans = sqrt(`Bare Ground`))
bg_log_assumptions <- sapply("bg_trans", function(x) {
  df <- dplyr::rename(bg, data = all_of(x))
  shapiro <- df %>% group_by(Site) %>% shapiro_test(data)
  levene <- df %>% levene_test(data ~ Site)
  return(list(shapiro = shapiro, levene = levene))
}, simplify = FALSE, USE.NAMES = TRUE)

# Not coded but I tried all data transformations. Instead, use nonparametric test
litter_analysis <- litter %>% t_test(Litter ~ Site, var.equal = TRUE) %>% add_significance()
litter_effects <- litter %>% cohens_d(Litter ~ Site, var.equal = TRUE)

bg_analysis <- litter %>% rstatix::wilcox_test(`Bare Ground` ~ Site) %>% add_significance()
bg_effects <- litter %>% rstatix::wilcox_effsize(`Bare Ground` ~ Site)

lit_bg_analysis <- bind_rows(litter_analysis, bg_analysis)
lit_bg_effects <- rbind(litter_effects, bg_effects)

# Invaded and pristine sites have significantly different litter and bare ground 
# covers. I haven't built code to show these in a graph because I think it might
# be a bit useless, but I can if it becomes not so useless.

# Write outputs
write.csv(litter_normal_analysis, file.path(
  results_dir, "02a_litter_bg_norm_assumption.csv"), row.names = TRUE)
write.csv(litter_var_eq_analysis, file.path(
  results_dir, "02b_litter_bg_var_eq_assumption.csv"), row.names = TRUE)
write.csv(lit_bg_analysis, file.path(
  results_dir, "02c_litter_bg_ttest.csv"), row.names = FALSE)
write.csv(lit_bg_effects, file.path(
  results_dir, "02d_litter_bg_effects.csv"), row.names = FALSE)

```

Next, there were ground temperature recordings taken for Jordann's project in 2017. There are two thermochrons that were within my invaded area, and then I will use data from two data loggers that were in pristine areas close to my area of invasion.

```{r Ground temperature analysis}

temperature_dir <- file.path(data_dir, "Temperature Data 2017")
layer_dir <- file.path("./03_Field_Surveys/layers")
dir.create(layer_dir, showWarnings = FALSE)

# Load the study area boundary
aoi <- st_read("./01_Spatial_Data/layers/study_area/aoi.gpkg", quiet = TRUE) %>% 
  st_geometry()

# Load the thermochron location points by unqipping the kmz and loading the 
# subsequent kml
unzip(file.path(data_dir, "Temperature Data 2017", "LDB Knapweed Study Sites.kmz"), 
      exdir = tempdir(), overwrite = TRUE)

temp_sites <- st_read(file.path(tempdir(), "doc.kml"), quiet = TRUE) %>% 
  st_transform(3005)

# Get the temperature stations within the study area
sites_within <- st_set_agr(temp_sites, "constant") %>% 
  st_intersection(aoi) %>% 
  dplyr::mutate(Type = "Invaded")

# Sites 1, 2, 12, 17, and 18 all have no recorded SK (aka these are pristine sites),
# so filter to those sites only and take the nearest two to the AOI polygon 
sites_near <- data.frame(Name = temp_sites$Name, dist = as.numeric(st_distance(aoi, temp_sites))) %>% 
  dplyr::arrange(dist) %>% 
  dplyr::filter(Name %in% paste("Site", c(1, 2, 12, 17, 18))) %>% 
  dplyr::slice(c(1, 2)) %>% 
  merge(temp_sites, by = "Name", all.x = TRUE) %>% 
  st_sf(sf_column_name = "geometry") %>% 
  st_zm() %>% 
  dplyr::mutate(Type = "Pristine")

# Combine the filtered points - these are the weather stations that will be used
site_list <- bind_rows(sites_near, sites_within) %>% 
  dplyr::select(-Description) %>% 
  replace_na(list(dist = 0))

# Now, read CSV files matching the selected sites
temp_data <- sapply(unique(site_list$Type), function(x) {
  df <- dplyr::filter(site_list, Type == x)
  bind_rows(lapply(unique(df$Name), function(y) {
    read.csv(file.path(temperature_dir, paste0(tolower(y), ".csv")), skip = 7, header = TRUE) %>% 
      dplyr::select(-X) %>% 
      drop_na() %>% 
      dplyr::mutate(Date = as.Date(Date...Time.1, format = "%m.%d.%Y"), 
                    DateTime = as.POSIXct(strptime(paste(as.Date(Date...Time.1, format = "%m.%d.%Y"), 
                                              Date...Time.2), format = "%Y-%m-%d %H:%M:%S")), 
                    month = month.abb[as.numeric(strftime(Date, "%m"))], 
                    day = as.numeric(strftime(Date, "%d"))) %>% 
      dplyr::select(-c(Date...Time, Date...Time.1)) %>% 
      dplyr::rename(time = Date...Time.2, temp = Temperature...C.)
  })) %>% group_by(DateTime, Date, month, day, time) %>% 
    dplyr::summarise(temp = mean(temp), .groups = "drop") %>% 
    dplyr::mutate(Type = as.factor(x))
}, simplify = FALSE, USE.NAMES = TRUE)

temp_data_df <- do.call(rbind, temp_data)
temp_sum_stats <- temp_data_df %>% 
  group_by(Type) %>% 
  get_summary_stats(temp)

# Test for normality
temp_shapiro <- temp_data_df %>% 
  group_by(Type) %>% 
  shapiro_test(temp) %>% mutate(metric = "temp_raw")

# Data is not normal and it doesn't make a lot of sense to transform temp data.
# When the data is plotted, it is all over the place; however, the smoothed lines
# appear to be more normalized and perhaps they are. I'll create the plot for the
# temperature data and then extract the smooth line data to see if that data is 
# normal enough for a proper t-test instead of the Mann Whitney test done below:
temp_analysis <- temp_data_df %>% rstatix::wilcox_test(temp ~ Type) %>% 
  add_significance() %>% mutate(metric = "temp_raw")
temp_effect <- temp_data_df %>% wilcox_effsize(temp ~ Type) %>% mutate(metric = "temp_raw")

# The data are indeed different across the entire range of when data was collected
# Now, create figure
p <- ggplot(temp_data_df, aes(DateTime, temp)) +
  geom_line(aes(color = Type), size = 1, alpha = 0.25) +
  geom_smooth(
    aes(fill = Type, color = Type),
    size = 1,
    alpha = 0.6) +
  scale_color_manual(values = c("grey33", "black"), name = "Survey Site") +
  scale_fill_manual(values = c("grey33", "black"), name = "Survey Site") +
  scale_y_continuous(
    expand = c(0, 0),
    limits = c(5, 40),
    breaks = c(5, 10, 15, 20, 25, 30, 35, 40)) +
  scale_x_datetime(
    expand = c(0, 0),
    limits = as.POSIXct(c(
      "2017-06-12 00:00:00", "2017-09-03 00:00:00")),
    breaks = as.POSIXct(c(
        "2017-06-15 00:00:00",
        "2017-07-01 00:00:00",
        "2017-07-15 00:00:00",
        "2017-08-01 00:00:00",
        "2017-08-15 00:00:00",
        "2017-09-01 00:00:00")),
    date_labels = "%b. %d") +
  xlab("Date") +
  ylab("Temperature (\u00B0C)") +
  theme_classic(base_size = 14) +
  theme(
    text = element_text(family = "Times New Roman"),
    legend.position = "right",
    legend.title = element_text(face = "bold", size = 16),
    legend.background = element_rect(
      fill = "white",
      size = 1,
      linetype = "solid",
      color = "black"),
    axis.title = element_text(face = "bold", size = 16),
    axis.title.y = element_text(margin = margin(t = 0, 10, 0, 0)),
    axis.text = element_text(color = "black"),
    axis.line = element_line(size = 1)) +
  ggsave(file.path(plot_dir, "ground_temps.png"), width = 6, height = 3, dpi = 300)

# There may be something to those smoothed lines that the noise has a hard
# time picking up
smooth_data <- ggplot_build(p)$data[[2]] %>% 
  dplyr::mutate(Type = as.factor(ifelse(fill == "black", "Invaded", "Pristine")))

# Test of equal variances and normality

temp_smooth_shapiro <- smooth_data %>% 
  group_by(Type) %>% 
  shapiro_test(y) %>% mutate(metric = "temp_smooth")

# The smoothed data are not normal, proceed with Mann Whitney test
temp_smooth_analysis <- smooth_data %>% rstatix::wilcox_test(y ~ Type) %>% 
  add_significance() %>% mutate(metric = "temp_smooth")
temp_smooth_effect <- smooth_data %>% wilcox_effsize(y ~ Type) %>% 
  mutate(metric = "temp_smooth")

# Again, these smoothed values show significant differences
# Prep outputs
temp_shapiro <- rbind(temp_shapiro, temp_smooth_shapiro)
temp_ttest <- rbind(temp_analysis, temp_smooth_analysis)
temp_effect <- rbind(temp_effect, temp_smooth_effect)

# Write outputs
st_write(temp_sites, file.path(
  layer_dir, "all_thermochron_locations.gpkg"), delete_layer = TRUE, quiet = TRUE)
st_write(site_list, file.path(
  layer_dir, "thermochron_analysis_locations.gpkg"), delete_layer = TRUE, quiet = TRUE)
write.csv(temp_shapiro, file.path(
  results_dir, "03a_temperature_norm_assumptions.csv"), row.names = FALSE)
write.csv(temp_ttest, file.path(
  results_dir, "03b_temperature_ttests.csv"), row.names = FALSE)
write.csv(temp_effect, file.path(
  results_dir, "03c_temperature_effects.csv"), row.names = FALSE)

```
